Adult Straying

Calculate the proportion of adults straying to non-natal streams to spawn

adult_stray(
  wild,
  natal_flow,
  south_delta_watershed,
  cross_channel_gates_closed,
  prop_bay_trans = 0,
  prop_delta_trans = 0,
  .intercept = fallRunDSM::params$.adult_stray_intercept,
  .wild = fallRunDSM::params$.adult_stray_wild,
  .natal_flow = fallRunDSM::params$.adult_stray_natal_flow,
 
    .cross_channel_gates_closed = fallRunDSM::params$.adult_stray_cross_channel_gates_closed,
  .prop_bay_trans = fallRunDSM::params$.adult_stray_prop_bay_trans,
  .prop_delta_trans = fallRunDSM::params$.adult_stray_prop_delta_trans
)

Source

IP-117068

Arguments

wild

Variable indicator of wild fish returning

natal_flow

Variable describing proportion flows at tributary junctions coming from natal watershed in October

south_delta_watershed

Variable indicator if watershed feeds into South Delta

cross_channel_gates_closed

Variable describing number of days gates are closed for each month

prop_bay_trans

Variable describing proportion transport to the bay

prop_delta_trans

Variable describing proportion transport to the delta

.intercept

Intercept

.wild

Coefficient for wild variable

.natal_flow

Coefficient for natal_flow variable

.cross_channel_gates_closed

Coefficient for cross_channel_gates_closed variable

.prop_bay_trans

Coefficient for prop_bay_trans variable

.prop_delta_trans

Coefficient for prop_delta_trans variable

Details

See params for details on parameter sources