Fall Run Chinook Model

Fall Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

fall_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = fallRunDSM::params,
  stochastic = FALSE
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels

Source

IP-117068

Examples

fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = fall_run_seeds)

#> $spawners
#>                                 1      2      3      4      5      6      7
#> Upper Sacramento River      25822  26400  21506  18851  20670  24978  25305
#> Antelope Creek                233    252    176    127    195    330    339
#> Battle Creek                37533  37558  37334  37184  37274  37517  37549
#> Bear Creek                    202    212    150    108    180    313    310
#> Big Chico Creek               207    218    160    121    181    306    312
#> Butte Creek                  1097   1127    771    455    585    960    893
#> Clear Creek                 10301  10820   6728   3984   4910   6830   5968
#> Cottonwood Creek             3516   3521   3011   2633   3531   5165   5167
#> Cow Creek                    2014   2026   1351   1004   1241   1729   1670
#> Deer Creek                    523    537    422    348    406    534    540
#> Elder Creek                   207    218    153    113    209    376    370
#> Mill Creek                    549    586    387    249    295    415    419
#> Paynes Creek                  243    251    170    125    209    362    348
#> Stony Creek                   203    213    152    112    191    336    336
#> Thomes Creek                  203    213    152    112    191    336    339
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    200    209    144    101    163    282    279
#> Feather River              135781 134262 130720 129237 132740 139991 139153
#> Yuba River                  25890  25181  23398  22326  26548  33828  33655
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              42515  45866  37263  32021  35148  46490  50445
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River               240    252    189    146    195    302    302
#> Cosumnes River                218    232    154     97    144    247    250
#> Mokelumne River              7503   7515   7481   7476   7537   7736   7752
#> Merced River                 3708   3731   3769   3810   3806   3838   3875
#> Stanislaus River             4593   4741   5045   4926   4655   4576   4703
#> Tuolumne River               1458   1493   1468   1427   1429   1512   1551
#> San Joaquin River               0      0      0      0      0      0      0
#>                                 8      9     10     11     12     13     14
#> Upper Sacramento River      21727  20348  19914  18956  18845  19055  17962
#> Antelope Creek                226    203    199    143    109    135    105
#> Battle Creek                37336  37234  37239  37184  37171  37205  37166
#> Bear Creek                    190    163    163    116     83    108     85
#> Big Chico Creek               208    190    188    128     98    125     98
#> Butte Creek                   619    548    515    367    290    331    277
#> Clear Creek                  4155   3957   3980   3406   3156   3308   3011
#> Cottonwood Creek             4139   3518   3201   2725   2659   2927   2544
#> Cow Creek                    1198   1049   1004    874    852    911    827
#> Deer Creek                    423    391    392    352    324    346    328
#> Elder Creek                   241    218    218    142     99    127     97
#> Mill Creek                    305    280    279    234    206    229    208
#> Paynes Creek                  215    190    185    128     99    122     98
#> Stony Creek                   227    200    190    128     93    117     91
#> Thomes Creek                  227    204    196    133    103    140    101
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    177    154    153    109     82    105     87
#> Feather River              133697 131013 130180 128605 128877 130703 129889
#> Yuba River                  28424  26678  27112  23998  21803  22609  21973
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              38662  35283  39481  33481  29519  30808  29923
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River               198    183    194    148    112    134    118
#> Cosumnes River                156    137    143    102     73     95     79
#> Mokelumne River              7522   7444   7475   7389   7309   7322   7311
#> Merced River                 3855   3868   3867   3722   3600   3617   3655
#> Stanislaus River             4845   4827   4681   4351   4088   4098   4221
#> Tuolumne River               1487   1451   1435   1365   1308   1330   1324
#> San Joaquin River               0      0      0      0      0      0      0
#>                                15     16     17     18     19     20
#> Upper Sacramento River      16943  20096  23802  24631  25433  22516
#> Antelope Creek                 81    163    257    296    392    303
#> Battle Creek                37133  37291  37453  37459  37490  37350
#> Bear Creek                     70    139    218    248    336    262
#> Big Chico Creek                81    149    229    286    395    295
#> Butte Creek                   239    385    542    586    711    584
#> Clear Creek                  2744   3625   4607   4814   5207   4419
#> Cottonwood Creek             2190   3406   4648   4830   5626   4868
#> Cow Creek                     752   1135   1536   1597   1783   1454
#> Deer Creek                    316    389    474    499    573    496
#> Elder Creek                    73    167    270    310    424    335
#> Mill Creek                    195    260    334    369    458    382
#> Paynes Creek                   81    154    240    274    367    288
#> Stony Creek                    73    149    234    262    359    293
#> Thomes Creek                   77    156    246    270    362    295
#> Upper-mid Sacramento River      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0
#> Bear River                     74    136    205    238    326    262
#> Feather River              128923 133886 138807 139929 142208 137530
#> Yuba River                  21621  31140  39537  41588  48043  41220
#> Lower-mid Sacramento River      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0
#> American River              29351  30876  32223  34953  41212  40408
#> Lower Sacramento River          0      0      0      0      0      0
#> Calaveras River               111    160    217    232    301    254
#> Cosumnes River                 72    120    177    193    262    213
#> Mokelumne River              7308   7337   7372   7391   7498   7548
#> Merced River                 3703   3717   3725   3741   3832   3838
#> Stanislaus River             4354   4363   4337   4356   4502   4489
#> Tuolumne River               1328   1373   1424   1438   1514   1485
#> San Joaquin River               0      0      0      0      0      0
#> 
#> $juvenile_biomass
#>                                    1         2          3          4          5
#> Upper Sacramento River      283228.0  447638.4  1618387.4  2879223.0  3088472.6
#> Antelope Creek                5086.8    7881.4    21006.6    36078.6    39030.2
#> Battle Creek                 24115.2   37020.6   259112.8   496936.4   552828.4
#> Bear Creek                    3454.0    5118.2    19279.6    30081.2    32185.0
#> Big Chico Creek               4584.4    6311.4    15700.0    26721.4    29327.6
#> Butte Creek                  13313.6   20943.8   137626.2   183501.6   227273.2
#> Clear Creek                 122334.4  203943.0   994281.0  1310573.2  1423581.8
#> Cottonwood Creek             53568.4   88987.6   638550.4   944417.8  1223689.4
#> Cow Creek                    40035.0   57933.0   199704.0   310357.6   328381.2
#> Deer Creek                    3140.0    3893.6    19499.4    36800.8    37994.0
#> Elder Creek                   3611.0    6280.0    32970.0    48450.2    57713.2
#> Mill Creek                    4333.2    5809.0    15009.2    23644.2    25120.0
#> Paynes Creek                  4804.2    7127.8    29484.6    43112.2    45875.4
#> Stony Creek                   3893.6    5903.2    25088.6    36267.0    45247.4
#> Thomes Creek                  3862.2    5809.0    25088.6    36267.0    45655.6
#> Upper-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Sutter Bypass                    0.0       0.0        0.0        0.0        0.0
#> Bear River                    3077.2    5149.6    29830.0    38967.4    50522.6
#> Feather River              1847105.0 2666990.4 11606790.2 17034751.2 20745100.8
#> Yuba River                  291235.0  457560.8  3591971.6  5157104.6  6739696.0
#> Lower-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Yolo Bypass                      0.0       0.0        0.0        0.0        0.0
#> American River             5609829.8 6844948.8 20113427.0 38265798.4 42496257.6
#> Lower Sacramento River           0.0       0.0        0.0        0.0        0.0
#> Calaveras River               8603.6    8603.6    16956.0    21634.6    21760.2
#> Cosumnes River                1884.0    1884.0     3925.0     5055.4     5620.6
#> Mokelumne River              52218.2   52218.2   136150.4   204790.8   230224.8
#> Merced River                109774.4  157816.4   212546.6   254591.2   371744.6
#> Stanislaus River            191131.8  319275.2   417463.0   496245.6   759346.2
#> Tuolumne River               46283.6   57838.8    72691.0    89678.4   151191.0
#> San Joaquin River                0.0       0.0        0.0        0.0        0.0
#>                                     6          7          8          9
#> Upper Sacramento River      3375625.6  3732989.0  3903051.4  3997377.0
#> Antelope Creek                45624.2    52186.8    55326.8    56771.2
#> Battle Creek                 576975.0   659996.6   673121.8   682887.2
#> Bear Creek                    36047.2    40254.8    41699.2    42327.2
#> Big Chico Creek               34791.2    41824.8    43520.4    44525.2
#> Butte Creek                  243412.8   287718.2   292177.0   295474.0
#> Clear Creek                 1518880.8  1741444.0  1782703.6  1806096.6
#> Cottonwood Creek            1301090.4  1492724.6  1528363.6  1555681.6
#> Cow Creek                    359090.4   385937.4   397555.4   402799.2
#> Deer Creek                    39752.4    42986.6    43646.0    44148.4
#> Elder Creek                   64213.0    79661.8    82111.0    83649.6
#> Mill Creek                    27537.8    30112.6    31243.0    31871.0
#> Paynes Creek                  52029.8    57556.2    59879.8    60853.2
#> Stony Creek                   51339.0    59283.2    62046.4    63396.6
#> Thomes Creek                  51558.8    62140.6    64621.2    65940.0
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                    53568.4    60131.0    61198.6    62486.0
#> Feather River              21664806.8 25099621.4 25751768.0 26490547.2
#> Yuba River                  6935412.2  8829271.8  8955970.8  9119125.2
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             43723809.2 58097724.4 59100483.4 60586048.8
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River               21760.2    29955.6    29955.6    29955.6
#> Cosumnes River                 5620.6     7567.4     7567.4     7567.4
#> Mokelumne River              230224.8   275440.8   275440.8   275440.8
#> Merced River                 421419.4   543910.8   569282.0   576472.6
#> Stanislaus River             880895.6  1061665.4  1131844.4  1150496.0
#> Tuolumne River               162871.8   197474.6   203409.2   204916.4
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    10         11         12         13
#> Upper Sacramento River      4332697.6  4380519.8  4415656.4  4442566.2
#> Antelope Creek                63773.4    64370.0    64746.8    65092.2
#> Battle Creek                 752846.4   759032.2   768954.6   775705.6
#> Bear Creek                    46346.4    46691.8    46943.0    47131.4
#> Big Chico Creek               49580.6    50020.2    50302.8    50585.4
#> Butte Creek                  313277.8   314345.4   316103.8   317234.2
#> Clear Creek                 1924286.2  1937034.6  1948338.6  1957319.0
#> Cottonwood Creek            1729103.8  1742700.0  1752245.6  1762953.0
#> Cow Creek                    436554.2   439443.0   442237.6   444121.6
#> Deer Creek                    46911.6    47131.4    47351.2    47508.2
#> Elder Creek                   91562.4    92033.4    92410.2    92881.2
#> Mill Creek                    34885.4    35199.4    35482.0    35639.0
#> Paynes Creek                  65783.0    66254.0    66536.6    66787.8
#> Stony Creek                   68389.2    68860.2    69142.8    69394.0
#> Thomes Creek                  76616.0    77212.6    77872.0    78657.0
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                    69268.4    70022.0    70555.8    71811.8
#> Feather River              29369550.4 29920306.4 30653056.8 32176365.0
#> Yuba River                  9636032.0  9848892.6  9973990.2 10401187.2
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             64598529.2 65677527.4 67217603.2 68533326.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River               29955.6    29955.6    29955.6    29955.6
#> Cosumnes River                 7598.8     7598.8     7598.8     7598.8
#> Mokelumne River              275440.8   275440.8   275440.8   275440.8
#> Merced River                 598546.8   608846.0   686466.8   701852.8
#> Stanislaus River            1208774.4  1231099.8  1388476.6  1420661.6
#> Tuolumne River               210285.8   212483.8   229125.8   232485.6
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    14         15         16         17
#> Upper Sacramento River      5900970.6  5987414.8  7672527.2  8148865.2
#> Antelope Creek                83555.4    84277.6   113982.0   121926.2
#> Battle Creek                1063141.2  1073440.4  1366779.2  1410174.0
#> Bear Creek                    58937.8    59157.6    79096.6    84434.6
#> Big Chico Creek               62486.0    62988.4    95173.4    99977.6
#> Butte Creek                  369452.4   370771.2   436522.8   458973.8
#> Clear Creek                 2382380.8  2404957.4  2897152.4  3056884.2
#> Cottonwood Creek            2354874.4  2368282.2  3040933.0  3346360.8
#> Cow Creek                    614655.0   618611.4   808958.2   855901.2
#> Deer Creek                    72848.0    73256.2   106885.6   111281.6
#> Elder Creek                  116808.0   117216.2   154017.0   168398.2
#> Mill Creek                    44713.6    45184.6    64275.8    68263.6
#> Paynes Creek                  81451.6    81954.0   106540.2   113165.6
#> Stony Creek                   84937.0    85439.4   108141.6   118880.4
#> Thomes Creek                  96806.2    97371.4   119979.4   131503.2
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                    93132.4    93697.6   137563.4   152509.8
#> Feather River              40472433.4 41029532.2 51391218.2 56183517.6
#> Yuba River                 15965549.8 16091212.6 23196907.0 25600451.4
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             74738562.6 75863310.6 87516855.4 97404118.8
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River               30458.0    30458.0    31086.0    35105.2
#> Cosumnes River                 7850.0     7850.0     9137.4    10267.8
#> Mokelumne River              275440.8   275440.8   282600.0   325147.0
#> Merced River                 749863.4   759220.6   816557.0   870156.8
#> Stanislaus River            1531409.4  1554708.2  1678612.6  1747127.4
#> Tuolumne River               243381.4   245391.0   258233.6   272489.2
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    18          19          20
#> Upper Sacramento River      8249784.8   9698706.4   9963377.0
#> Antelope Creek               123810.2    161396.0    165697.8
#> Battle Creek                1420567.4   1586045.4   1604948.2
#> Bear Creek                    85502.2    110842.0    114986.8
#> Big Chico Creek              102018.6    137029.6    140326.6
#> Butte Creek                  462239.4    510438.4    541524.4
#> Clear Creek                 3070103.6   3490298.4   3620043.2
#> Cottonwood Creek            3376347.8   3724919.2   3879375.8
#> Cow Creek                    865101.4   1046091.0   1072058.8
#> Deer Creek                   112600.4    131691.6    134988.6
#> Elder Creek                  170659.0    208558.8    214210.8
#> Mill Creek                    69142.8     86852.4     88328.2
#> Paynes Creek                 114610.0    146198.4    151034.0
#> Stony Creek                  120324.8    153326.2    163405.6
#> Thomes Creek                 133041.8    167236.4    177378.6
#> Upper-mid Sacramento River        0.0         0.0         0.0
#> Sutter Bypass                     0.0         0.0         0.0
#> Bear River                   157471.0    181021.0    198793.4
#> Feather River              57029308.0  62212694.4  64942108.0
#> Yuba River                 26309118.0  29079885.4  31368034.8
#> Lower-mid Sacramento River        0.0         0.0         0.0
#> Yolo Bypass                       0.0         0.0         0.0
#> American River             99747877.6 112077495.8 122745017.8
#> Lower Sacramento River            0.0         0.0         0.0
#> Calaveras River               35230.8     46220.8     62360.4
#> Cosumnes River                10927.2     14004.4     17646.8
#> Mokelumne River              346593.2    417274.6    496873.6
#> Merced River                 941497.6   1024896.0   1183183.4
#> Stanislaus River            1863684.2   1987274.6   2385301.0
#> Tuolumne River               302727.4    319243.8    421105.4
#> San Joaquin River                 0.0         0.0         0.0
#> 
#> $proportion_natural
#>                                     1          2           3           4
#> Upper Sacramento River     0.36019673 0.37420455 0.231795778 0.123600870
#> Antelope Creek             0.80686695 0.82142857 0.744318182 0.645669291
#> Battle Creek               0.01135001 0.01200809 0.006080249 0.002070783
#> Bear Creek                 0.81683168 0.82547170 0.753333333 0.657407407
#> Big Chico Creek            0.78260870 0.79357798 0.718750000 0.628099174
#> Butte Creek                0.82224248 0.82697427 0.747081712 0.571428571
#> Clear Creek                0.74662654 0.75878004 0.612068966 0.344879518
#> Cottonwood Creek           0.41069397 0.41153082 0.311856526 0.213064945
#> Cow Creek                  0.65342602 0.65547878 0.483345670 0.304780876
#> Deer Creek                 0.45697897 0.47113594 0.327014218 0.183908046
#> Elder Creek                0.82125604 0.83027523 0.758169935 0.672566372
#> Mill Creek                 0.70491803 0.72354949 0.581395349 0.349397590
#> Paynes Creek               0.81481481 0.82071713 0.735294118 0.640000000
#> Stony Creek                0.81773399 0.82629108 0.756578947 0.669642857
#> Thomes Creek               0.81773399 0.82629108 0.756578947 0.669642857
#> Upper-mid Sacramento River         NA         NA          NA          NA
#> Sutter Bypass                      NA         NA          NA          NA
#> Bear River                 0.81500000 0.82296651 0.743055556 0.633663366
#> Feather River              0.08191131 0.07152433 0.046366279 0.035423292
#> Yuba River                 0.22618772 0.20440014 0.143772972 0.102660575
#> Lower-mid Sacramento River         NA         NA          NA          NA
#> Yolo Bypass                        NA         NA          NA          NA
#> American River             0.40999647 0.45310252 0.326838956 0.216639081
#> Lower Sacramento River             NA         NA          NA          NA
#> Calaveras River            0.67916667 0.69444444 0.592592593 0.472602740
#> Cosumnes River             0.82568807 0.83620690 0.753246753 0.608247423
#> Mokelumne River            0.02878848 0.03033932 0.025932362 0.025280899
#> Merced River               0.04989213 0.05574913 0.065269302 0.075328084
#> Stanislaus River           0.16655781 0.19257541 0.241228940 0.222898904
#> Tuolumne River             0.13443073 0.15472204 0.140326975 0.115627190
#> San Joaquin River                  NA         NA          NA          NA
#>                                      5          6          7          8
#> Upper Sacramento River     0.200725689 0.33857795 0.34712507 0.23960970
#> Antelope Creek             0.769230769 0.86363636 0.86725664 0.80088496
#> Battle Creek               0.004480335 0.01092838 0.01177129 0.00613349
#> Bear Creek                 0.794444444 0.88178914 0.88064516 0.80526316
#> Big Chico Creek            0.751381215 0.85294118 0.85576923 0.78365385
#> Butte Creek                0.666666667 0.79687500 0.78163494 0.68497577
#> Clear Creek                0.468431772 0.61786237 0.56266756 0.37184116
#> Cottonwood Creek           0.413197395 0.59883833 0.59899361 0.49939599
#> Cow Creek                  0.437550363 0.59629844 0.58203593 0.41736227
#> Deer Creek                 0.300492611 0.46816479 0.47407407 0.32860520
#> Elder Creek                0.822966507 0.90159574 0.90000000 0.84647303
#> Mill Creek                 0.450847458 0.60963855 0.61336516 0.46885246
#> Paynes Creek               0.784688995 0.87569061 0.87068966 0.79069767
#> Stony Creek                0.806282723 0.88988095 0.88988095 0.83700441
#> Thomes Creek               0.806282723 0.88988095 0.89085546 0.83700441
#> Upper-mid Sacramento River          NA         NA         NA         NA
#> Sutter Bypass                       NA         NA         NA         NA
#> Bear River                 0.773006135 0.86879433 0.86738351 0.79096045
#> Feather River              0.060878409 0.10952133 0.10415873 0.06760062
#> Yuba River                 0.245366883 0.40776871 0.40472441 0.29517309
#> Lower-mid Sacramento River          NA         NA         NA         NA
#> Yolo Bypass                         NA         NA         NA         NA
#> American River             0.286332081 0.46044311 0.50274556 0.35119756
#> Lower Sacramento River              NA         NA         NA         NA
#> Calaveras River            0.605128205 0.74503311 0.74503311 0.61111111
#> Cosumnes River             0.736111111 0.84615385 0.84800000 0.75641026
#> Mokelumne River            0.033169696 0.05804033 0.05998452 0.03124169
#> Merced River               0.074356280 0.08207400 0.09083871 0.08612192
#> Stanislaus River           0.177658432 0.16346154 0.18605146 0.20990712
#> Tuolumne River             0.116864941 0.16534392 0.18633140 0.15131137
#> San Joaquin River                   NA         NA         NA         NA
#>                                      9          10          11          12
#> Upper Sacramento River     0.188077452 0.170382645 0.128455370 0.123321836
#> Antelope Creek             0.778325123 0.773869347 0.685314685 0.587155963
#> Battle Creek               0.003410861 0.003544671 0.002070783 0.001721772
#> Bear Creek                 0.773006135 0.773006135 0.681034483 0.554216867
#> Big Chico Creek            0.763157895 0.760638298 0.648437500 0.540816327
#> Butte Creek                0.644160584 0.621359223 0.468664850 0.327586207
#> Clear Creek                0.340409401 0.344221106 0.233705226 0.173003802
#> Cottonwood Creek           0.411028994 0.352702281 0.239633028 0.220759684
#> Cow Creek                  0.334604385 0.304780876 0.201372998 0.180751174
#> Deer Creek                 0.273657289 0.275510204 0.193181818 0.123456790
#> Elder Creek                0.830275229 0.830275229 0.739436620 0.626262626
#> Mill Creek                 0.421428571 0.419354839 0.307692308 0.213592233
#> Paynes Creek               0.763157895 0.756756757 0.648437500 0.545454545
#> Stony Creek                0.815000000 0.805263158 0.710937500 0.602150538
#> Thomes Creek               0.818627451 0.811224490 0.721804511 0.640776699
#> Upper-mid Sacramento River          NA          NA          NA          NA
#> Sutter Bypass                       NA          NA          NA          NA
#> Bear River                 0.759740260 0.758169935 0.660550459 0.548780488
#> Feather River              0.048499004 0.042410509 0.030683099 0.032728881
#> Yuba River                 0.249044156 0.261065211 0.165180432 0.081135624
#> Lower-mid Sacramento River          NA          NA          NA          NA
#> Yolo Bypass                         NA          NA          NA          NA
#> American River             0.289062721 0.364656417 0.250798961 0.150242217
#> Lower Sacramento River              NA          NA          NA          NA
#> Calaveras River            0.579234973 0.603092784 0.479729730 0.312500000
#> Cosumnes River             0.722627737 0.734265734 0.627450980 0.479452055
#> Mokelumne River            0.021090811 0.025150502 0.013804304 0.003009988
#> Merced River               0.089193382 0.088957848 0.053465879 0.021388889
#> Stanislaus River           0.206960845 0.182226020 0.120202252 0.063600783
#> Tuolumne River             0.130254997 0.120557491 0.075457875 0.035168196
#> San Joaquin River                   NA          NA          NA          NA
#>                                     13          14           15          16
#> Upper Sacramento River     0.132983469 0.080224919 0.0249070413 0.177896099
#> Antelope Creek             0.666666667 0.571428571 0.4444444444 0.723926380
#> Battle Creek               0.002634055 0.001587472 0.0007001858 0.004934166
#> Bear Creek                 0.657407407 0.564705882 0.4714285714 0.733812950
#> Big Chico Creek            0.640000000 0.540816327 0.4444444444 0.697986577
#> Butte Creek                0.410876133 0.296028881 0.1841004184 0.493506494
#> Clear Creek                0.211003628 0.133178346 0.0488338192 0.280000000
#> Cottonwood Creek           0.292107960 0.185534591 0.0538812785 0.391661773
#> Cow Creek                  0.233809001 0.155985490 0.0718085106 0.385022026
#> Deer Creek                 0.179190751 0.134146341 0.1012658228 0.269922879
#> Elder Creek                0.708661417 0.618556701 0.4931506849 0.778443114
#> Mill Creek                 0.292576419 0.221153846 0.1692307692 0.376923077
#> Paynes Creek               0.631147541 0.540816327 0.4444444444 0.707792208
#> Stony Creek                0.683760684 0.593406593 0.4931506849 0.751677852
#> Thomes Creek               0.735714286 0.633663366 0.5194805195 0.762820513
#> Upper-mid Sacramento River          NA          NA           NA          NA
#> Sutter Bypass                       NA          NA           NA          NA
#> Bear River                 0.647619048 0.574712644 0.5000000000 0.727941176
#> Feather River              0.046242244 0.040265149 0.0330740054 0.068916840
#> Yuba River                 0.113892698 0.088244664 0.0734008603 0.356647399
#> Lower-mid Sacramento River          NA          NA           NA          NA
#> Yolo Bypass                         NA          NA           NA          NA
#> American River             0.185795897 0.161715069 0.1453783517 0.187589066
#> Lower Sacramento River              NA          NA           NA          NA
#> Calaveras River            0.425373134 0.347457627 0.3063063063 0.518750000
#> Cosumnes River             0.600000000 0.518987342 0.4722222222 0.683333333
#> Mokelumne River            0.004780115 0.003282725 0.0028735632 0.006814774
#> Merced River               0.025988388 0.036114911 0.0486092358 0.052192628
#> Stanislaus River           0.065885798 0.093105899 0.1208084520 0.122622049
#> Tuolumne River             0.051127820 0.046827795 0.0496987952 0.080844865
#> San Joaquin River                   NA          NA           NA          NA
#>                                     17          18         19          20
#> Upper Sacramento River     0.305898664 0.329259876 0.35041088 0.266255107
#> Antelope Creek             0.824902724 0.847972973 0.88520408 0.851485149
#> Battle Creek               0.009238245 0.009396941 0.01021606 0.006506024
#> Bear Creek                 0.830275229 0.850806452 0.88988095 0.858778626
#> Big Chico Creek            0.803493450 0.842657343 0.88607595 0.847457627
#> Butte Creek                0.640221402 0.667235495 0.72573840 0.666095890
#> Clear Creek                0.433470805 0.457831325 0.49875168 0.409368635
#> Cottonwood Creek           0.554216867 0.571014493 0.63170992 0.574363188
#> Cow Creek                  0.545572917 0.562930495 0.60852496 0.519944979
#> Deer Creek                 0.400843882 0.430861723 0.50436300 0.427419355
#> Elder Creek                0.862962963 0.880645161 0.91273585 0.889552239
#> Mill Creek                 0.514970060 0.560975610 0.64628821 0.575916230
#> Paynes Creek               0.812500000 0.835766423 0.87738420 0.843750000
#> Stony Creek                0.841880342 0.858778626 0.89693593 0.873720137
#> Thomes Creek               0.849593496 0.862962963 0.89779006 0.874576271
#> Upper-mid Sacramento River          NA          NA         NA          NA
#> Sutter Bypass                       NA          NA         NA          NA
#> Bear River                 0.819512195 0.844537815 0.88650307 0.858778626
#> Feather River              0.101925695 0.109126771 0.12340375 0.093586854
#> Yuba River                 0.493284771 0.518274502 0.58299856 0.513973799
#> Lower-mid Sacramento River          NA          NA         NA          NA
#> Yolo Bypass                         NA          NA         NA          NA
#> American River             0.221549825 0.282350585 0.39134233 0.379231835
#> Lower Sacramento River              NA          NA         NA          NA
#> Calaveras River            0.645161290 0.668103448 0.74418605 0.696850394
#> Cosumnes River             0.785310734 0.803108808 0.85496183 0.821596244
#> Mokelumne River            0.011530114 0.014071168 0.02814084 0.034578696
#> Merced River               0.054228188 0.058273189 0.08063674 0.082073997
#> Stanislaus River           0.117362232 0.121212121 0.14971124 0.147248830
#> Tuolumne River             0.113764045 0.122392211 0.16644650 0.150168350
#> San Joaquin River                   NA          NA         NA          NA
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    0    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    0
#>