Filling Natal Tributary and Regional Habitat

Allocates juvenile fish onto the floodplain and in-channel habitat. Once available habitat is full, fish are assigned to out migrate. The fill_natal function is used to route fish within their natal tributary and the fill_regional function is used to route fish on the mainstem, bypasses, and deltas.

fill_natal(
  juveniles,
  inchannel_habitat,
  floodplain_habitat,
  territory_size = fallRunDSM::params$territory_size,
  up_to_size_class = 2
)

fill_regional(
  juveniles,
  habitat,
  floodplain_habitat = NULL,
  territory_size = fallRunDSM::params$territory_size,
  up_to_size_class = 3
)

Source

IP-117068

Arguments

juveniles

An n by 4 matrix of juvenile fish size s, m, l, vl

inchannel_habitat

The available tributary habitat in square meters

floodplain_habitat

The available floodplain habitat in square meters (NULL for bypasses and delta)

territory_size

A length 4 array of juvenile fish territory requirements s, m, l, vl

up_to_size_class

Which size class and under is moved to floodplain and inchannel habitat (small = 1, medium = 2, large = 3, very large = 4)

Details

See params for details on parameter sources