Get Migratory Survival Rates

Calculates the juvenile out migration survival rates in all regions for a month and year of the simulation

get_migratory_survival(
  year,
  month,
  cc_gates_prop_days_closed,
  freeport_flows,
  vernalis_flows,
  stockton_flows,
  vernalis_temps,
  prisoners_point_temps,
  CVP_exports,
  SWP_exports,
  upper_sacramento_flows,
  delta_inflow,
  avg_temp_delta,
  avg_temp,
  delta_proportion_diverted,
  ..surv_juv_outmigration_sj_int,
  .surv_juv_outmigration_san_joaquin_medium,
  .surv_juv_outmigration_san_joaquin_large,
  min_survival_rate,
  stochastic
)

Source

IP-117068

Arguments

year

The simulation year, 1-20

month

The simulation month, 1-8

cc_gates_prop_days_closed

The number of days and proportion of days the Delta Cross Channel gates are typically closed for each month. More details at delta_cross_channel_closed

freeport_flows

Flow in cms at Freeport (CALSIM node C400). More details at freeport_flow

vernalis_flows

Flow in cms at Vernalis (CALSIM node C639). More details at vernalis_flow

stockton_flows

Flow in cms at Stockton (CALSIM node C417A). More details at stockton_flow

vernalis_temps

Monthly water temperature (°C) at Vernalis from 1980-2000. More details at vernalis_temperature

prisoners_point_temps

Monthly water temperature (°C) at Prisoner's Point from 1980-2000. More details at prisoners_point_temperature

CVP_exports

Total exports for CVP in cms. Value is obtained using CALSIM variable DEL_CVP_EXP. More details at cvp_exports

SWP_exports

Total exports for SWP in cms. Value is obtained using CALSIM variable DEL_SWP_EXP. More details at swp_exports

upper_sacramento_flows

Average monthly flows on the Upper Sacramento River in cubic meters per second, more details at upper_sacramento_flows

delta_inflow

Variable describing delta inflow in cubic meters per second, more details at delta_inflow

avg_temp_delta

Variable describing monthly mean temperature in celsius, more details at delta_temperature

avg_temp

Variable describing monthly mean temperature in celsius, more details at stream_temperature

delta_proportion_diverted

Variable describing diversions from the delta in cubic meters per second, more details at delta_proportion_diverted

..surv_juv_outmigration_sj_int

Intercept for surv_juv_outmigration_san_joaquin

.surv_juv_outmigration_san_joaquin_medium

Size related intercept for surv_juv_outmigration_san_joaquin medium sized fish

.surv_juv_outmigration_san_joaquin_large

Size related intercept for surv_juv_outmigration_san_joaquin large sized fish

min_survival_rate

Estimated survival rate if temperature threshold is exceeded

stochastic

TRUE FALSE value indicating if model is being run stochastically