Get Rearing Survival Rates

Calculates the juvenile inchannel, floodplain, bypasses, and deltas rearing survival rates for a month and year of the simulation

get_rearing_survival(
  year,
  month,
  survival_adjustment,
  mode,
  avg_temp,
  avg_temp_delta,
  prob_strand_early,
  prob_strand_late,
  proportion_diverted,
  total_diverted,
  delta_proportion_diverted,
  delta_total_diverted,
  weeks_flooded,
  prop_high_predation,
  contact_points,
  delta_contact_points,
  delta_prop_high_predation,
  ..surv_juv_rear_int,
  .surv_juv_rear_contact_points,
  ..surv_juv_rear_contact_points,
  .surv_juv_rear_prop_diversions,
  ..surv_juv_rear_prop_diversions,
  .surv_juv_rear_total_diversions,
  ..surv_juv_rear_total_diversions,
  ..surv_juv_bypass_int,
  ..surv_juv_delta_int,
  .surv_juv_delta_contact_points,
  ..surv_juv_delta_contact_points,
  .surv_juv_delta_total_diverted,
  ..surv_juv_delta_total_diverted,
  .surv_juv_rear_avg_temp_thresh,
  .surv_juv_rear_high_predation,
  .surv_juv_rear_stranded,
  .surv_juv_rear_medium,
  .surv_juv_rear_large,
  .surv_juv_rear_floodplain,
  .surv_juv_bypass_avg_temp_thresh,
  .surv_juv_bypass_high_predation,
  .surv_juv_bypass_medium,
  .surv_juv_bypass_large,
  .surv_juv_bypass_floodplain,
  .surv_juv_delta_avg_temp_thresh,
  .surv_juv_delta_high_predation,
  .surv_juv_delta_prop_diverted,
  .surv_juv_delta_medium,
  .surv_juv_delta_large,
  min_survival_rate,
  stochastic
)

Arguments

year
The simulation year, 1-20

month
The simulation month, 1-8

mode
The mode that the model is being run in

avg_temp
More details at stream_tempetature

avg_temp_delta
More details at delta_temprature

prob_strand_early
More details at prop_strand_early

prob_strand_late
More details at prop_strand_late

proportion_diverted
More details at proportion_diverted

total_diverted
More details at total_diverted

delta_proportion_diverted
More details at delta_proportion_diverted

delta_total_diverted
More details at delta_total_diverted

weeks_flooded
More details at weeks_flooded

prop_high_predation
More details at prop_high_predation

contact_points
More details at contact_points

delta_contact_points
More details at delta_contact_points

delta_prop_high_predation
More details at delta_prop_high_predation

..surv_juv_rear_int
Intercept for surv_juv_rear

.surv_juv_rear_contact_points
Coefficient for surv_juv_rear contact_points variable

..surv_juv_rear_contact_points
Calibrated coefficient for surv_juv_rear contact_points variable

.surv_juv_rear_prop_diversions
Coefficient for surv_juv_rear prop_diversions variable

..surv_juv_rear_prop_diversions
Calibrated coefficient for surv_juv_rear prop_diversions variable

.surv_juv_rear_total_diversions
Coefficient for surv_juv_rear total_diversions variable

..surv_juv_rear_total_diversions
Calibrated coefficient for surv_juv_rear total_diversions variable

..surv_juv_bypass_int
Intercept for surv_juv_bypass

..surv_juv_delta_int
Intercept for surv_juv_delta

.surv_juv_delta_contact_points
Coefficient for surv_juv_delta contact_points variable

..surv_juv_delta_contact_points
Calibrated coefficient for surv_juv_delta contact_points variable

.surv_juv_delta_total_diverted
Coefficient for surv_juv_delta total_diversions variable

..surv_juv_delta_total_diverted
Calibrated coefficient for surv_juv_delta total_diversions variable

.surv_juv_rear_avg_temp_thresh
Coefficient for surv_juv_rear avg_temp_thresh variable

.surv_juv_rear_high_predation
Coefficient for surv_juv_rear high_predation variable

.surv_juv_rear_stranded
Coefficient for surv_juv_rear stranded variable

.surv_juv_rear_medium
Size related intercept for surv_juv_rear medium sized fish

.surv_juv_rear_large
Size related intercept for surv_juv_rear large sized fish

.surv_juv_rear_floodplain
Additional intercept for surv_juv_rear floodplain rearing benefit

.surv_juv_bypass_avg_temp_thresh
Coefficient for surv_juv_bypass avg_temp_thresh variable

.surv_juv_bypass_high_predation
Coefficient for surv_juv_bypass high_predation variable

.surv_juv_bypass_medium
Size related intercept for surv_juv_bypass medium sized fish

.surv_juv_bypass_large
Size related intercept for surv_juv_bypass large sized fish

.surv_juv_bypass_floodplain
Additional intercept for surv_juv_bypass floodplain rearing benefit

.surv_juv_delta_avg_temp_thresh
Coefficient for surv_juv_delta avg_temp_thresh variable

.surv_juv_delta_high_predation
Coefficient for surv_juv_delta high_predation variable

.surv_juv_delta_prop_diverted
Coefficient for surv_juv_delta prop_diversions variable

.surv_juv_delta_medium
Size related intercept for surv_juv_delta medium sized fish

.surv_juv_delta_large
Size related intercept for surv_juv_delta large sized fish

min_survival_rate
estimated survival rate if temperature threshold is exceeded

stochastic
TRUE FALSE value indicating if model is being run stochastically

Source

IP-117068