Get Spawning Adults

Get Spawning Adults

get_spawning_adults(
  year,
  adults,
  hatch_adults,
  mode,
  month_return_proportions,
  prop_flow_natal,
  south_delta_routed_watersheds,
  cc_gates_days_closed,
  gates_overtopped,
  tisdale_bypass_watershed,
  yolo_bypass_watershed,
  migratory_temperature_proportion_over_20,
  natural_adult_removal_rate,
  cross_channel_stray_rate,
  stray_rate,
  ..surv_adult_enroute_int,
  .adult_stray_intercept,
  .adult_stray_wild,
  .adult_stray_natal_flow,
  .adult_stray_cross_channel_gates_closed,
  .adult_stray_prop_bay_trans,
  .adult_stray_prop_delta_trans,
  .adult_en_route_migratory_temp,
  .adult_en_route_bypass_overtopped,
  .adult_en_route_adult_harvest_rate,
  stochastic
)

Arguments

year
the year of simulation

adults
potential spawning adults for each watershed (length = 31) values must be integer

hatch_adults
total hatchery adults

mode
a value meant to be inherited to determine if model is in "seed", "calibrate", or "simulate" mode

month_return_proportions
The proportion of fish returning for each month

prop_flow_natal
More details at proportion_flow_natal

south_delta_routed_watersheds
More details at south_delta_routed_watersheds

cc_gates_days_closed
More details at delta_cross_channel_closed

gates_overtopped
More details at gates_overtopped

tisdale_bypass_watershed
More details at tisdale_bypass_watershed

yolo_bypass_watershed
More details at yolo_bypass_watershed

migratory_temperature_proportion_over_20
More details at migratory_temperature_proportion_over_20

natural_adult_removal_rate
More details at natural_adult_removal_rate

cross_channel_stray_rate
More details at cross_channel_stray_rate

stray_rate
More details at stray_rate

..surv_adult_enroute_int
Intercept for surv_adult_enroute

.adult_stray_intercept
Intercept for adult_stray

.adult_stray_wild
Coefficient for adult_stray wild variable

.adult_stray_natal_flow
Coefficient for adult_stray natal_flow variable

.adult_stray_cross_channel_gates_closed
Coefficient for adult_stray cross_channel_gates_closed variable

.adult_stray_prop_bay_trans
Coefficient for adult_stray prop_bay_trans variable

.adult_stray_prop_delta_trans
Coefficient for adult_stray prop_delta_trans variable

.adult_en_route_migratory_temp
Coefficient for surv_adult_enroute migratory_temp variable

.adult_en_route_bypass_overtopped
Coefficient for surv_adult_enroute bypass_overtopped variable

.adult_en_route_adult_harvest_rate
Adult harvest rate for surv_adult_enroute

stochastic
TRUE FALSE value indicating if model is being run stochastically

Source

IP-117068

Details

See params for details on parameter sources