Model Parameters

A list containing all parameters needed for running the fall_run_model. Parameters with a single period in front are model or submodel coefficients. Parameters with double periods in front of them are calibrated model parameters.

params

params_2019

Habitat Inputs

spawning_habitat: More details at fr_spawn

inchannel_habitat_fry: More details at fr_fry

inchannel_habitat_juvenile: More details at fr_juv

floodplain_habitat: More details at fr_fp

sutter_habitat: More details at sutter_habitat

yolo_habitat: More details at yolo_habitat

delta_habitat: More details at delta_habitat

spawn_decay_rate: More details at spawn_decay_rate

rear_decay_rate: More details at rear_decay_rate

Spawning Adults

prop_flow_natal: More details at proportion_flow_natal

south_delta_routed_watersheds: More details at south_delta_routed_watersheds

cc_gates_days_closed: More details at delta_cross_channel_closed

cross_channel_stray_rate: More details at cross_channel_stray_rate

stray_rate: More details at stray_rate

gates_overtopped: More details at gates_overtopped

tisdale_bypass_watershed: More details at tisdale_bypass_watershed

yolo_bypass_watershed: More details at yolo_bypass_watershed

migratory_temperature_proportion_over_20: More details at migratory_temperature_proportion_over_20

..surv_adult_enroute_int: Intercept, source: calibration

.adult_stray_intercept: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District

.adult_stray_wild: Coefficient for wild variable, source: Estimated with coded wire tag data 2010–2013 (Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)

.adult_stray_natal_flow: Coefficient for natal_flow variable, source: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District

.adult_stray_cross_channel_gates_closed: Coefficient for cross_channel_gates_closed variable, source: Empirical model fit using 2008–2011 tagging data provided by East Bay Municipal Utility District.

.adult_stray_prop_bay_trans: Coefficient for prop_bay_trans variable, source: Estimated with coded wire tag data 2010–2013 (Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)

.adult_stray_prop_delta_trans: Coefficient for prop_delta_trans variable, source: Estimated with coded wire tag data 2010–2013 (Kormos et al. 2012, Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)

.adult_en_route_migratory_temp: Coefficient for migratory_temp variable, source: Schreck et al. (1994)

.adult_en_route_bypass_overtopped: Coefficient for bypass_overtopped variable, source: Expert opinion Ted Sommer, California Department of Water Resources (tributaries above bypasses only)

.adult_en_route_adult_harvest_rate: Adult harvest rate, source: (Palmer-Zwahlen & Kormos 2013-2015, Palmer-Zwahlen et al. 2018)

degree_days: More details at degree_days

month_return_proportions: More details at month_return_proportions

..surv_adult_prespawn_int: Intercept, source: Calibration Estimate

.adult_prespawn_deg_day: Coefficient for deg_day variable, source: Colvin et al. (2018)

prob_nest_scoured: More details at prob_nest_scoured

spawn_success_sex_ratio: Variable describing the female to male spawning ratio, default 0.5, source: expert opinion from SIT members.

spawn_success_redd_size: Variable describing the size of redds including defensible space, default value 9.29 square meters, source: expert opinion from SIT members.

spawn_success_fecundity: Variable describing the number of eggs per female, default value 5522, source: Moyle, P. B. 2002. Inland Fishes of California. University of California Press, Berkeley CA

hatchery_allocation: More details at hatchery_allocation

natural_adult_removal_rate: More details at natural_adult_removal_rate

Egg to Fry Survival

proportion_hatchery: More details at proportion_hatchery

prob_nest_scoured: More details at prob_nest_scoured

mean_egg_temp_effect: More details at egg_temperature_effect

.surv_egg_to_fry_proportion_natural: Coefficient for proportion_natural variable, source: Chilcote et al. (2011)

.surv_egg_to_fry_scour: Coefficient for scour variable, source: Montgomery et al. (1996)

..surv_egg_to_fry_int: Intercept, source: Calibration

Tributary Rearing Survival

avg_temp: More details at stream_temperature

prob_strand_early: More details at prob_strand_early

prob_strand_late: More details at prob_strand_late

proportion_diverted: More details at proportion_diverted

total_diverted: More details at total_diverted

weeks_flooded: More details at weeks_flooded

prop_high_predation: More details at prop_high_predation

contact_points: More details at contact_points

..surv_juv_rear_int: Intercept, source: calibration (varies by tributary)

.surv_juv_rear_contact_points: Coefficient for contact_points variable, source: inherited from previous calibration

..surv_juv_rear_contact_points: Coefficient for contact_points variable, source: calibration

.surv_juv_rear_prop_diversions: Coefficient for prop_diversions variable, source: Newman and Brandes (2010)

..surv_juv_rear_prop_diversions: Coefficient for prop_diversions variable, source: calibration

.surv_juv_rear_total_diversions: Coefficient for total_diversions variable, source: inherited from previous calibration

..surv_juv_rear_total_diversions: Coefficient for total_diversions variable, source: calibration

.surv_juv_rear_avg_temp_thresh: Coefficient for avg_temp_thresh variable, source: Runge et al (2008)

.surv_juv_rear_high_predation: Coefficient for high_predation variable, source: Cavallo et al. (2012)

.surv_juv_rear_stranded: Coefficient for stranded variable, source: USFWS (2006) and CDWR (2006)

.surv_juv_rear_medium: Size related intercept for medium sized fish, source: Connor et al. (2004)

.surv_juv_rear_large: Size related intercept for large sized fish, source: Connor et al. (2004)

.surv_juv_rear_floodplain: Additional intercept for floodplain rearing benefit, source: Sommer et al. (2001)

Bypass Rearing Survival

avg_temp: See Tributary Rearing Survival avg_temp above

prop_high_predation: See Tributary Rearing Survival prop_high_predation above

..surv_juv_bypass_int: Intercept, source: calibration

.surv_juv_bypass_avg_temp_thresh: Coefficient for avg_temp_thresh variable, source: Marine and Chech (2004)

.surv_juv_bypass_high_predation: Coefficient for high_predation variable, source: Cavallo et al. (2012)

.surv_juv_bypass_medium: Size related intercept for medium sized fish, source: Connor et al. (2004)

.surv_juv_bypass_large: Size related intercept for large sized fish, source: Connor et al. (2004)

.surv_juv_bypass_floodplain: Additional intercept for floodplain rearing benefit, source: Sommer et al. (2001)

Delta Rearing Survival

avg_temp_delta: More details at delta_temperature

delta_proportion_diverted: More details at delta_proportion_diverted

delta_total_diverted: More details at delta_total_diverted

delta_contact_points: More details at delta_contact_points

delta_prop_high_predation: More details at delta_prop_high_predation

..surv_juv_delta_int: Intercept, source: calibration

.surv_juv_delta_contact_points: Coefficient for contact_points variable, source: inherited from previous calibration

..surv_juv_delta_contact_points: Coefficient for contact_points variable, source: calibration

.surv_juv_delta_total_diverted: Coefficient for total_diversions variable, source: inherited from previous calibration

..surv_juv_delta_total_diverted: Coefficient for total_diversions variable, source: calibration

.surv_juv_delta_avg_temp_thresh: Coefficient for avg_temp_thresh variable, source: Marine and Chech (2004)

.surv_juv_delta_high_predation: Coefficient for high_predation variable, source: Cavallo et al. (2012)

.surv_juv_delta_prop_diverted: Coefficient for prop_diversions variable, source: Newman and Brandes (2010)

.surv_juv_delta_medium: Size related intercept for medium sized fish, source: Connor et al. (2004)

.surv_juv_delta_large: Size related intercept for large sized fish, source: Connor et al. (2004)

Tributary Migratory Survival

cc_gates_prop_days_closed: More details at delta_cross_channel_closed

upper_sacramento_flows: More details at upper_sacramento_flows

freeport_flows: More details at freeport_flow

vernalis_flows: More details at vernalis_flow

stockton_flows: More details at stockton_flow

vernalis_temps: More details at vernalis_temperature

prisoners_point_temps: More details at prisoners_point_temperature

CVP_exports: More details at cvp_exports

SWP_exports: More details at swp_exports

..surv_juv_outmigration_sj_int Intercept, source: calibration

.surv_juv_outmigration_san_joquin_medium Size related intercept for medium sized fish, source: Connor et al. (2004)

.surv_juv_outmigration_san_joaquin_large Size related intercept for large sized fish, source: Connor et al. (2004)

min_survival_rate Estimated survival rate if temperature threshold is exceeded, source: expert opinion

avg_temp: More details at stream_temperature

Delta Migratory Survival

delta_inflow: More details at delta_inflow

avg_temp_delta: More details at delta_temperature

delta_proportion_diverted: More details at delta_proportion_diverted

.surv_juv_outmigration_sac_delta_intercept_one Intercept for model one, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_intercept_two Intercept for model two, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_intercept_three Intercept for model three, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_delta_flow Coefficient for delta_flow variable, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_avg_temp Coefficient for avg_temp variable, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_perc_diversions Coefficient for perc_diversions variable, source: this parameter and function have been deprecated and its documentation is for archival purposes.

.surv_juv_outmigration_sac_delta_medium Size related intercept for medium sized fish, source: Connor et al. (2004)

.surv_juv_outmigration_sac_delta_large Size related intercept for large sized fish, source: Connor et al. (2004)

Delta Routing and Rearing

freeport_flows: More details at freeport_flow

cc_gates_days_closed: More details at delta_cross_channel_closed

growth_rates: More details at: growth_rates_inchannel

Tributary Routing

prop_pulse_flows: More details at proportion_pulse_flows

proportion_flow_bypass: More details at proportion_flow_bypasses

.pulse_movement_intercept: Intercept, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_proportion_pulse: Coefficient for proportion_pulse variable, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_medium: Size related intercept for medium sized fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_large: Size related intercept for large sized fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_vlarge: Size related intercept for very large sized fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_medium_pulse: Additional coefficient for proportion_pulse variable for medium size fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_large_pulse: Additional coefficient for proportion_pulse variable for large size fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

.pulse_movement_very_large_pulse: Additional coefficient for proportion_pulse variable for very large size fish, source: Empirical model fit to CVPIA Chinook salmon screw trap abundance estimates from American River, Stanislaus River, Feather River, Mokelumne River, Sacramento River, Tuolumne River, Clear Creek 2008–2015.

territory_size: More details at territory_by_size

Tributary Rearing

growth_rates: More details at: growth_rates_inchannel

growth_rates_floodplain: More details at: growth_rates_floodplain

weeks_flooded: More details at: weeks_flooded

Ocean Entry Success

.ocean_entry_success_length: Size related intercept representing the fork lengths for each size classes, source: Satterthwaite et al. (2014)

..ocean_entry_success_int: Intercept, source: Calibration (Varies by tributary)

.ocean_entry_success_months: Coefficient for month variable, source: Satterthwaite et al. (2014)

Generating Model Outputs

mass_by_size_class: More details at mass_by_size_class

diversity_group: More details at diversity_group