Juvenile Pulse Movement

Calculates the proportion of juveniles in each size class that migrate due to pulse flow

pulse_movement(
  proportion_pulse,
  .intercept = fallRunDSM::params$.pulse_movement_intercept,
  .proportion_pulse = fallRunDSM::params$.pulse_movement_proportion_pulse,
  .medium = fallRunDSM::params$.pulse_movement_medium,
  .large = fallRunDSM::params$.pulse_movement_large,
  .vlarge = fallRunDSM::params$.pulse_movement_vlarge,
  .medium_pulse = fallRunDSM::params$.pulse_movement_medium_pulse,
  .large_pulse = fallRunDSM::params$.pulse_movement_large_pulse,
  .very_large_pulse = fallRunDSM::params$.pulse_movement_very_large_pulse
)

Source

IP-117068

Arguments

proportion_pulse

The proportion of flow that is a pulse, estimated by the standard deviation of flow divided by the median flow for each month

.intercept

Intercept

.proportion_pulse

Coefficient for the proportion_pulse variable

.medium

Size related intercept for medium sized fish

.large

Size related intercept for large sized fish

.vlarge

Size related intercept for very large sized fish

.medium_pulse

Additional coefficient for proportion_pulse variable for medium size fish

.large_pulse

Additional coefficient for proportion_pulse variable for large size fish

.very_large_pulse

Additional coefficient for proportion_pulse variable for very large size fish

Details

See params for details on parameter sources