Route Natal Streams

Determines if juveniles stay in their natal tributary, are detoured to a bypass, or out migrate during a simulated month

route(
  year,
  month,
  juveniles,
  inchannel_habitat,
  floodplain_habitat,
  prop_pulse_flows,
  .pulse_movement_intercept,
  .pulse_movement_proportion_pulse,
  .pulse_movement_medium,
  .pulse_movement_large,
  .pulse_movement_vlarge,
  .pulse_movement_medium_pulse,
  .pulse_movement_large_pulse,
  .pulse_movement_very_large_pulse,
  territory_size,
  stochastic
)

Arguments

year
The current simulation year, 1-20

month
The current simulation month, 1-8

juveniles
An n by 4 matrix of juvenile fish by watershed and size class

inchannel_habitat
A vector of available habitat in square meters

floodplain_habitat
A vector of available floodplain habitat in square meters

prop_pulse_flows
The proportion of pulse flows

.pulse_movement_intercept
Intercept for pulse_movement

.pulse_movement_proportion_pulse
Coefficient for pulse_movement proportion_pulse variable

.pulse_movement_medium
Size related intercept for pulse_movement medium sized fish

.pulse_movement_large
Size related intercept for pulse_movement large sized fish

.pulse_movement_vlarge
Size related intercept for pulse_movement very large sized fish

.pulse_movement_medium_pulse
Additional coefficient for pulse_movement proportion_pulse variable for medium size fish

.pulse_movement_large_pulse
Additional coefficient for pulse_movement proportion_pulse variable for large size fish

.pulse_movement_very_large_pulse
Additional coefficient for pulse_movement proportion_pulse variable for very large size fish

territory_size
Array of juvenile fish territory requirements for fill_natal

stochastic
TRUE FALSE value indicating if model is being run stochastically

Source

IP-117068

Details

See params for details on parameter sources