Route and Rear in the Deltas

Determines if juveniles stay in the delta or out migrate to golden gate during a simulated month. Then the remaining juveniles in the delta rear (growth and survival rates applied) and survival rates are applied to out migrating juveniles

route_and_rear_deltas(
  year,
  month,
  migrants,
  north_delta_fish,
  south_delta_fish,
  north_delta_habitat,
  south_delta_habitat,
  freeport_flows,
  cc_gates_days_closed,
  rearing_survival_delta,
  migratory_survival_delta,
  migratory_survival_bay_delta,
  juveniles_at_chipps,
  growth_rates,
  location_index = c(rep(1, 24), 3, rep(2, 2), rep(4, 4)),
  territory_size,
  stochastic
)

Arguments

year
The current simulation year, 1-20

month
The current simulation month, 1-8

migrants
An n by 4 matrix of juvenile fish by watershed and size class

north_delta_fish
An n by 4 matrix of juvenile fish by watershed and size class

south_delta_fish
An n by 4 matrix of juvenile fish by watershed and size class

north_delta_habitat
A vector of available habitat in square meters

south_delta_habitat
A vector of available habitat in square meters

rearing_survival_delta
The rearing survival rate for North and South Delta

migratory_survival_delta
The outmigration survival rate for North and South Delta

migratory_survival_bay_delta
The outmigration survival rate in the Bay Delta

juveniles_at_chipps
The accumulated juveniles at Chipps Island for the current year

growth_rates
The delta growth rate

location_index
Migratory survival probability location index for fish coming from 4 areas (1-4) representing "northern_fish", "cosumnes_mokelumne_fish", "calaveras_fish", or "southern_fish" respectively

territory_size
Array of juvenile fish territory requirements for fill_natal

stochastic
TRUE FALSE value indicating if model is being run stochastically

freeport_flow
Monthly mean flow at freeport in cubic feet per second

cc_gates_closed
Number of days the Cross Channel gates are closed during the month

Source

IP-117068