Route Regions

Determines if juveniles stay in the region (Sections of Mainstem Sacramento River or San Joaquin River) or out migrate during a simulated month

route_regional(
  month,
  year,
  migrants,
  inchannel_habitat,
  floodplain_habitat,
  prop_pulse_flows,
  migration_survival_rate,
  proportion_flow_bypass,
  detour = NULL,
  territory_size,
  stochastic
)

Source

IP-117068

Arguments

month

The current simulation month, 1-8

year

The current simulation year, 1-20

migrants

An n by 4 matrix of juvenile fish by watershed and size class

inchannel_habitat

A vector of available habitat in square meters

floodplain_habitat

A vector of available floodplain habitat in square meters

prop_pulse_flows

The proportion of pulse flows

migration_survival_rate

The outmigration survival rate

proportion_flow_bypass

Variable describing the proportion of flows routed through the bypasses, more details at proportion_flow_bypasses

detour

Values can be 'sutter' or 'yolo' if some juveniles are detoured on to that bypass, otherwise NULL

territory_size

Array of juvenile fish territory requirements for fill_regional

stochastic

TRUE FALSE value indicating if model is being run stochastically