Route To South Delta

Entrains northern juveniles into the South Delta

route_to_south_delta(
  freeport_flow,
  dcc_closed,
  month,
  mean_freeport_flow = 21546.19,
  sd_freeport_flow = 14375.9,
  .sss_int = 1.892235,
  .sss_freeport_discharge = 2.170375,
  .sss_upper_asymptote = 0.3512465,
  .dcc_intercept = -1.48962,
  .dcc_freeport_discharge = -1.248865,
  .gs_intercept = -2.948145,
  .gs_freeport_discharge = -2.911835,
  .gs_dcc_effect_on_routing = -0.554843,
  .gs_lower_asymptote = 0.2729845
)

Arguments

freeport_flow

Monthly mean flow at freeport in cubic feet per second

dcc_closed

Number of days the Delta Cross Channel gates are closed during the month

month

Current simulation month as an integer for calculating number of days the Delta Cross Channel gates are open

mean_freeport_flow

Mean of flow at freeport for standardizing discharge

sd_freeport_flow

Standard Deviation of flow at freeport for standardizing discharge

.sss_int

Intercept for Sutter and Steamboat Sloughs, source: This submodel is adapted from Perry et al. (2018)

.sss_freeport_discharge

Coefficient for freeport_flow for Sutter and Steamboat Sloughs, source: This submodel is adapted from Perry et al. (2018)

.sss_upper_asymptote

Parameter representing the upper asymptote for Sutter and Steamboat Sloughs, source: This submodel is adapted from Perry et al. (2018)

.dcc_intercept

Intercept for Delta Cross Channel, source: This submodel is adapted from Perry et al. (2018)

.dcc_freeport_discharge

Coefficient for freeport_flow for Delta Cross Channel Gates, source: This submodel is adapted from Perry et al. (2018)

.gs_intercept

Intercept for Georgiana Slough, source: This submodel is adapted from Perry et al. (2018)

.gs_freeport_discharge

Coefficient for freeport_flow for Georgiana Slough, source: This submodel is adapted from Perry et al. (2018)

.gs_dcc_effect_on_routing

Parameter representing the dcc effect on routing, source: This submodel is adapted from Perry et al. (2018)

.gs_lower_asymptote

Parameter representing the lower asymptote for Georgiana Slough, source: This submodel is adapted from Perry et al. (2018)