Egg to Fry Survival

Calculates the survival of transitioning from egg to fry

surv_egg_to_fry(
  proportion_natural,
  scour,
  temperature_effect = fallRunDSM::params$mean_egg_temp_effect,
  ..surv_egg_to_fry_int = fallRunDSM::params$..surv_egg_to_fry_int,
  .proportion_natural = fallRunDSM::params$.surv_egg_to_fry_proportion_natural,
  .scour = fallRunDSM::params$.surv_egg_to_fry_scour
)

Source

IP-117068

Arguments

proportion_natural

Variable describing the proportion of natural-origin spawners

scour

Variable describing the probability of redd scouring event

temperature_effect

Variable describing the effect of inchannel temperature on egg survival. Fall and spring estimated by C. Hammersmark (CBEC Ecoengineering Inc.). Winter-run value was calibrated.

..surv_egg_to_fry_int

Intercept

.proportion_natural

Coefficient for proportion_natural variable

.scour

Coefficient for scour variable

Details

See params for details on parameter sources