Juvenile Delta Survival

Calculates the juvenile rearing survival in the deltas

surv_juv_delta(
  avg_temp,
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diverted,
  total_diverted,
  ..surv_juv_delta_int = fallRunDSM::params$..surv_juv_delta_int,
  .avg_temp_thresh = fallRunDSM::params$.surv_juv_delta_avg_temp_thresh,
  .high_predation = fallRunDSM::params$.surv_juv_delta_high_predation,
  .surv_juv_delta_contact_points = fallRunDSM::params$.surv_juv_delta_contact_points,
  ..surv_juv_delta_contact_points = fallRunDSM::params$..surv_juv_delta_contact_points,
  .prop_diverted = fallRunDSM::params$.surv_juv_delta_prop_diverted,
  .surv_juv_delta_total_diverted = fallRunDSM::params$.surv_juv_delta_total_diverted,
  ..surv_juv_delta_total_diverted = fallRunDSM::params$..surv_juv_delta_total_diverted,
  .medium = fallRunDSM::params$.surv_juv_delta_medium,
  .large = fallRunDSM::params$.surv_juv_delta_large,
  min_survival_rate = fallRunDSM::params$min_survival_rate
)

Arguments

avg_temp
Variable representing average temperature in the delta

max_temp_thresh
Variable representing the probability of exceeding the max temperature

avg_temp_thresh
Variable representing the probability of exceeding the average temperature

high_predation
Variable representing an indicator for high predation in delta

contact_points
Variable representing the number of contact points in watershed

prop_diverted
Variable representing the proportion of water diverted

total_diverted
Variable representing the total diversions

..surv_juv_delta_int
intercept, source: calibration

.avg_temp_thresh
Coefficient for avg_temp_thresh variable

.high_predation
Coefficient for high_predation variable

.surv_juv_delta_contact_points
Coefficient for contact_points variable

..surv_juv_delta_contact_points
Calibrated coefficient for contact_points variable

.prop_diverted
Coefficient for prop_diversions variable

.surv_juv_delta_total_diverted
Coefficient for total_diversions variable

..surv_juv_delta_total_diverted
Calibrated coefficient for total_diversions variable

.medium
size related intercept for medium sized fish

.large
size related intercept for large sized fish

min_survival_rate
estimated survival rate if temperature threshold is exceeded

Source

IP-117068

Details

See params for details on parameter sources