Juvenile Delta Outmigration Survival

Calculates the North and South Delta juvenile out migration survival

surv_juv_outmigration_delta(
  prop_DCC_closed,
  hor_barr,
  freeport_flow,
  vernalis_flow,
  stockton_flow,
  vernalis_temperature,
  prisoners_point_temperature,
  CVP_exp,
  SWP_exp,
  trap_trans
)

Arguments

prop_DCC_closed
proportion of days the Delta Cross Channel Gates are closed

freeport_flow
average daily discharge at Freeport in cubic meters per second

vernalis_flow
average daily discharge at Vernalis in cubic meters per second

stockton_flow
average daily discharge at Stockton in cubic meters per second

vernalis_temperature
average daily temperature at Vernalis in °C

prisoners_point_temperature
average daily temperature of the San Joaquin River at Prisoners Point °C

CVP_exp
average daily exports Central Valley Project in cubic meters per second

SWP_exp
average daily exports State Water Project in cubic meters per second

horr_barr
indicator if head of old river physical barrier in place

Trap_trans
proportion of smolts trapped at Vernalis and transported to Chips island

Source

IP-117068

Details

function returns proportion of fish from the Sacramento at Feeeport (northern_fish) Mokelumne and Cosumnes (cosumnes_mokelumne_fish), Calaveras (calaveras_fish) and San Joaquin tributaries from Vernalis (southern_fish) arriving alive at Chipps Island in four size groups (35-42mm, 42-72mm, 72-110mm, >110mm). Note that the models were fit to data that were >80 mm. Therefore, this does not predict outside of the data so sizes <= 80mm are assumed to me 80mm long as requested by Russ Perry.

Parameters

All parameters were derived from Perry et al. (2018)