Juvenile Rearing Survival

Calculates the juvenile rearing survival inchannel and on the floodplain

surv_juv_rear(
  max_temp_thresh,
  avg_temp_thresh,
  high_predation,
  contact_points,
  prop_diversions,
  total_diversions,
  stranded,
  weeks_flooded,
  ..surv_juv_rear_int = fallRunDSM::params$..surv_juv_rear_int,
  .avg_temp_thresh = fallRunDSM::params$.surv_juv_rear_avg_temp_thresh,
  .high_predation = fallRunDSM::params$.surv_juv_rear_high_predation,
  .surv_juv_rear_contact_points = fallRunDSM::params$.surv_juv_rear_contact_points,
  ..surv_juv_rear_contact_points = fallRunDSM::params$..surv_juv_rear_contact_points,
  .surv_juv_rear_prop_diversions = fallRunDSM::params$.surv_juv_rear_prop_diversions,
  ..surv_juv_rear_prop_diversions = fallRunDSM::params$..surv_juv_rear_prop_diversions,
  .surv_juv_rear_total_diversions = fallRunDSM::params$.surv_juv_rear_total_diversions,
 
    ..surv_juv_rear_total_diversions = fallRunDSM::params$..surv_juv_rear_total_diversions,
  .stranded = fallRunDSM::params$.surv_juv_rear_stranded,
  .medium = fallRunDSM::params$.surv_juv_rear_medium,
  .large = fallRunDSM::params$.surv_juv_rear_large,
  .floodplain = fallRunDSM::params$.surv_juv_rear_floodplain,
  min_survival_rate = fallRunDSM::params$min_survival_rate,
  stochastic
)

Source

IP-117068

Arguments

max_temp_thresh

Variable representing probability of exceeding the max temperature threshold

avg_temp_thresh

Variable representing probability of exceeding the avg temperature threshold

high_predation

Variable representing indicator of high predation for a watershed

contact_points

Variable representing total number of contact points per watershed

prop_diversions

Variable representing proportion of water diverted

total_diversions

Variable representing total amount of water diverted

stranded

Variable representing stranding rate per watershed

weeks_flooded

Variable representing total weeks flooded per watershed

..surv_juv_rear_int

Intercept, source: calibration (varies by tributary)

.avg_temp_thresh

Coefficient for avg_temp_thresh variable

.high_predation

Coefficient for high_predation variable

.surv_juv_rear_contact_points

Coefficient for contact_points variable

..surv_juv_rear_contact_points

Calibrated coefficient for contact_points variable

.surv_juv_rear_prop_diversions

Coefficient for prop_diversions variable

..surv_juv_rear_prop_diversions

Calibrated coefficient for prop_diversions variable

.surv_juv_rear_total_diversions

Coefficient for total_diversions variable

..surv_juv_rear_total_diversions

Calibrated coefficient for total_diversions variable

.stranded

Coefficient for stranded variable

.medium

Size related intercept for medium sized fish

.large

Size related intercept for large sized fish

.floodplain

Additional intercept for floodplain rearing benefit

min_survival_rate

Estimated survival rate if temperature threshold is exceeded

stochastic

TRUE FALSE value indicating if model is being run stochastically

Details

See params for details on parameter sources