Fall Run Chinook Model

Fall Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

fall_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = fallRunDSM::params,
  stochastic = FALSE
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels. Defaults to fallRunDSM::params.

stochastic
TRUE FALSE value indicating if model should be run stochastically. Defaults to FALSE.

Source

IP-117068

Examples

fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = fall_run_seeds)

#> $spawners
#>                                 1      2      3      4      5      6      7
#> Upper Sacramento River      60474  66570  40040  24133  36208  69273  76875
#> Antelope Creek                366    419    304    181    307    674    824
#> Battle Creek                38796  38791  38557  38420  38503  38734  38787
#> Bear Creek                    303    328    248    153    303    675    791
#> Big Chico Creek               305    332    251    148    271    588    703
#> Butte Creek                  1691   1824   1103    539    931   1781   1741
#> Clear Creek                 13377  12802   7110   4375   6064   9831   9519
#> Cottonwood Creek             6770   7042   4752   3184   5740  12067  13050
#> Cow Creek                    2711   3180   1964   1097   1493   2572   2784
#> Deer Creek                   1245   1389    875    511    795   1461   1563
#> Elder Creek                   315    345    272    193    477   1103   1254
#> Mill Creek                   1064   1172    718    369    473    791    876
#> Paynes Creek                  400    424    298    185    354    773    888
#> Stony Creek                   308    332    255    168    403    863    946
#> Thomes Creek                  306    331    252    160    376    798    886
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    303    334    291    222    385    820   1000
#> Feather River              181389 195180 165617 138793 151085 200527 219340
#> Yuba River                  30602  33257  31789  30241  32862  46807  52331
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              30671  31272  30381  29960  30729  33270  33882
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River               363    390    508    604    962   2029   2240
#> Cosumnes River               1615   1807   1503   1354   2090   4256   4856
#> Mokelumne River              7882   7914   7838   7771   7904   8329   8437
#> Merced River                 3917   4132   4598   4804   4257   4182   4819
#> Stanislaus River             5549   6819   7912   6868   6220   6522   7743
#> Tuolumne River               1590   1753   1872   1751   1583   1719   2035
#> San Joaquin River               0      0      0      0      0      0      0
#>                                 8      9     10     11     12     13     14
#> Upper Sacramento River      44892  29447  32518  26943  24333  26623  22445
#> Antelope Creek                437    257    263    178    125    143    103
#> Battle Creek                38553  38425  38422  38396  38391  38412  38391
#> Bear Creek                    393    218    226    149    100    119     85
#> Big Chico Creek               372    256    285    181    112    136     99
#> Butte Creek                   948    894   1117    672    305    350    284
#> Clear Creek                  5591   4036   4120   3590   3373   3590   3213
#> Cottonwood Creek             7046   3899   4119   3198   2645   2942   2573
#> Cow Creek                    1595   1008   1032    895    813    847    793
#> Deer Creek                    864    580    584    454    384    412    361
#> Elder Creek                   652    517    664    371    149    197    123
#> Mill Creek                    503    341    329    267    235    253    219
#> Paynes Creek                  448    264    267    170    118    136    100
#> Stony Creek                   527    348    321    203    129    148     98
#> Thomes Creek                  493    475    614    347    177    256    155
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    570    296    273    172     89    105     78
#> Feather River              178902 142276 136842 135812 136887 137850 133681
#> Yuba River                  37282  27234  26914  24758  22684  22550  21821
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              31340  30075  30331  29536  29772  30411  30099
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River              1003    898   1416    808    377    577    381
#> Cosumnes River               2822   1781   1816    989    311    273    180
#> Mokelumne River              7963   7643   7575   7557   7545   7555   7541
#> Merced River                 4735   4117   3798   3728   3673   3683   3679
#> Stanislaus River             8075   5767   4039   4013   3997   4013   3990
#> Tuolumne River               1919   1579   1390   1363   1347   1364   1341
#> San Joaquin River               0      0      0      0      0      0      0
#>                                15     16     17     18     19     20
#> Upper Sacramento River      18503  31256  45630  48531  55869  49002
#> Antelope Creek                 74    186    294    358    519    443
#> Battle Creek                38376  38533  38687  38821  38993  38750
#> Bear Creek                     64    162    252    312    465    405
#> Big Chico Creek                74    153    226    379    614    473
#> Butte Creek                   235    421    599    698    928    782
#> Clear Creek                  2880   4395   5973   5907   6320   5574
#> Cottonwood Creek             2235   5286   8324   9078  11466   9364
#> Cow Creek                     754   1232   1705   1852   2168   1725
#> Deer Creek                    325    538    746    886   1146    926
#> Elder Creek                    68    278    477    723   1179    950
#> Mill Creek                    195    283    367    439    596    531
#> Paynes Creek                   74    183    287    356    522    446
#> Stony Creek                    68    186    303    383    609    564
#> Thomes Creek                   71    229    375    434    654    603
#> Upper-mid Sacramento River      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0
#> Bear River                     65    132    193    231    411    466
#> Feather River              132068 141858 148376 141877 153697 168728
#> Yuba River                  21328  24080  26510  31224  40635  39577
#> Lower-mid Sacramento River      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0
#> American River              30252  30855  30961  31361  32949  32412
#> Lower Sacramento River          0      0      0      0      0      0
#> Calaveras River               144    172    217    298    529    614
#> Cosumnes River                150    201    259    333    487    453
#> Mokelumne River              7535   7571   7601   7604   7656   7701
#> Merced River                 3684   3715   3741   3746   3820   4111
#> Stanislaus River             3978   4035   4085   4091   4173   4709
#> Tuolumne River               1329   1386   1436   1442   1524   1616
#> San Joaquin River               0      0      0      0      0      0
#> 
#> $juvenile_biomass
#>                                    1         2          3          4         5
#> Upper Sacramento River      952770.2  377333.8  6506959.2  9327872.4  803714.4
#> Antelope Creek               11272.6    3202.8    33158.4    61261.4    5871.8
#> Battle Creek                 24052.4    7598.8   165195.4   202216.0   28605.4
#> Bear Creek                    8383.8    2198.0    41448.0    51527.4    5620.6
#> Big Chico Creek               5024.0    2135.2    25748.0    36738.0    5181.0
#> Butte Creek                  66034.2   12089.0   528713.2   269788.8  141080.2
#> Clear Creek                 232642.6   70304.6  1128264.8   851002.8  140766.2
#> Cottonwood Creek            114578.6   42358.6  1439470.2  1634966.6  273619.6
#> Cow Creek                    57587.6    9828.2   215875.0   269883.0   15291.8
#> Deer Creek                   60696.2   12999.6   280245.0   318804.2   20253.0
#> Elder Creek                  15480.2    6562.6   109711.6   115866.0   31117.4
#> Mill Creek                   30426.6    7410.4    72596.8    66756.4    8949.0
#> Paynes Creek                 12277.4    3485.4    54981.4    65280.6    7347.6
#> Stony Creek                  10079.4    4176.2    83210.0    55672.2   29484.6
#> Thomes Creek                  9074.6    3736.6    72596.8    47351.2   29233.4
#> Upper-mid Sacramento River       0.0       0.0        0.0        0.0       0.0
#> Sutter Bypass                    0.0       0.0        0.0        0.0       0.0
#> Bear River                   32342.0     659.4    93477.8    95393.2   54855.8
#> Feather River               782048.4  644924.6 15877724.0 24949749.2 7080857.0
#> Yuba River                 5585180.8  511914.2 11553159.0 15186641.4 4092299.2
#> Lower-mid Sacramento River       0.0       0.0        0.0        0.0       0.0
#> Yolo Bypass                      0.0       0.0        0.0        0.0       0.0
#> American River             3625946.4 1438120.0  6541938.8  6667507.4 3395282.0
#> Lower Sacramento River           0.0       0.0        0.0        0.0       0.0
#> Calaveras River             101579.0     785.0   226739.4   259521.0   20661.2
#> Cosumnes River              238043.4   20629.8   571637.0   558135.0  261467.8
#> Mokelumne River              93823.2     219.8   196532.6   183313.2   62172.0
#> Merced River               1033091.4    8164.0    41008.4   487893.2  855493.0
#> Stanislaus River            282725.6  446884.8   391746.4   376705.8 1230597.4
#> Tuolumne River               92849.8     125.6     1036.2    36204.2  132822.0
#> San Joaquin River                0.0       0.0        0.0        0.0       0.0
#>                                    6         7         8         9        10
#> Upper Sacramento River      489871.4 2375347.2  328506.8  136338.8 1291105.2
#> Antelope Creek                7598.8   22356.8    3171.4     942.0    7598.8
#> Battle Creek                  8132.6   37617.2    4867.0    3391.2   35890.2
#> Bear Creek                    5275.2   21414.8    1758.4     408.2    5306.6
#> Big Chico Creek               6091.6   32813.0    1632.8     690.8    6531.2
#> Butte Creek                  18306.2  482837.8    4427.4    4270.4   34351.6
#> Clear Creek                  62423.2  221464.2   36863.6   16390.8  146575.2
#> Cottonwood Creek             38747.6  417902.6   18463.2   15040.6  161584.4
#> Cow Creek                    11555.2   44902.0    3736.6    1130.4   14852.2
#> Deer Creek                   22639.4   92096.2    6342.8    2637.6   24303.6
#> Elder Creek                  10927.2  120544.6    3705.2    2292.2   21509.0
#> Mill Creek                    8572.2   23267.4    3516.8    1161.8    8729.2
#> Paynes Creek                  8635.0   25371.2    2857.4     785.0    6908.0
#> Stony Creek                  12497.2   35356.4    5024.0    2072.4    9357.2
#> Thomes Creek                  9388.6  110967.6    3893.6    2103.8   35984.4
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                     596.6   48732.8     973.4      94.2    2260.8
#> Feather River               547333.4 2399211.2  463746.6 1590943.8 2128480.4
#> Yuba River                  714789.6 3435568.2  574620.0  392751.2  630260.8
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River             1581115.6 4254888.4 1261840.4 1700247.2 3793622.4
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River              20064.6  257825.4   17175.8    2920.2   87386.2
#> Cosumnes River               66348.2  335791.6   12434.4   34163.2   17207.2
#> Mokelumne River                282.6    3202.8      62.8     345.4     157.0
#> Merced River                  8069.8   92975.4    4647.2      31.4    3202.8
#> Stanislaus River               690.8    1475.8     502.4       0.0      31.4
#> Tuolumne River                  31.4    2355.0      31.4       0.0       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                   11        12        13        14        15
#> Upper Sacramento River       69394.0  106728.6  120952.8 5596327.8  208684.4
#> Antelope Creek                 219.8     251.2     219.8   28605.4     314.0
#> Battle Creek                  2135.2   10738.8    5149.6  271421.6    4710.0
#> Bear Creek                     157.0     157.0     157.0   20567.0     125.6
#> Big Chico Creek                219.8     219.8     219.8   12277.4     219.8
#> Butte Creek                    910.6    1852.6    1413.0  144973.8     659.4
#> Clear Creek                   7630.2   18683.0   14161.4  699686.2   19059.8
#> Cottonwood Creek              7284.8    7159.2    9639.8 1432059.8    6719.6
#> Cow Creek                      565.2     942.0     659.4  202247.4     785.0
#> Deer Creek                     785.0    1036.2     785.0  183847.0    1821.2
#> Elder Creek                    471.0     439.6     628.0   73727.2     251.2
#> Mill Creek                     314.0     502.4     439.6   26407.4     785.0
#> Paynes Creek                   219.8     251.2     219.8   26470.2     251.2
#> Stony Creek                    376.8     345.4     314.0   32404.8     439.6
#> Thomes Creek                  1130.4     785.0    1287.4   49141.0     408.2
#> Upper-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Sutter Bypass                    0.0       0.0       0.0       0.0       0.0
#> Bear River                      31.4     125.6     565.2    8478.0       0.0
#> Feather River               705149.8   30803.4 1637290.2 6115432.6  312901.0
#> Yuba River                  267873.4   26344.6  386157.2 3628238.6  238608.6
#> Lower-mid Sacramento River       0.0       0.0       0.0       0.0       0.0
#> Yolo Bypass                      0.0       0.0       0.0       0.0       0.0
#> American River             1162239.6 3339515.6 1822927.0 5279847.2 1214175.2
#> Lower Sacramento River           0.0       0.0       0.0       0.0       0.0
#> Calaveras River              13910.2     282.6    1821.2     125.6      31.4
#> Cosumnes River                6123.0   14318.4    1161.8   16390.8    2951.6
#> Mokelumne River                125.6     847.8      94.2     533.8     345.4
#> Merced River                    31.4   26313.2    2323.6    7222.0      31.4
#> Stanislaus River                 0.0      94.2       0.0     125.6       0.0
#> Tuolumne River                   0.0      62.8       0.0      94.2       0.0
#> San Joaquin River                0.0       0.0       0.0       0.0       0.0
#>                                    16         17        18        19        20
#> Upper Sacramento River      5967538.6  3476011.4  492226.4 6451852.2 1891379.0
#> Antelope Creek                54981.4    12748.4    2480.6   83429.8    6845.2
#> Battle Creek                 514991.4    50805.2   12371.6  224478.6   25151.4
#> Bear Creek                    46126.6    10330.6    1884.0   63271.0    9043.2
#> Big Chico Creek               81671.4     5589.2    4490.2   38276.6    6311.4
#> Butte Creek                  247557.6    67133.2    8509.4  236410.6  158884.0
#> Clear Creek                  609034.4   293715.6   32404.8  519575.8  248499.6
#> Cottonwood Creek            1791495.6   753317.4   40160.6  864065.2  239833.2
#> Cow Creek                    270134.2    45812.6    9137.4  270793.6   28762.4
#> Deer Creek                   334221.6    57085.2   11712.2  360974.4   52595.0
#> Elder Creek                  188274.4    65877.2    9388.6  356609.8   26219.0
#> Mill Creek                    82425.0    15040.6    2920.2   89364.4    7096.4
#> Paynes Creek                  56080.4    13031.0    2386.4   62705.8    9231.6
#> Stony Creek                   66725.0    37303.2    3171.4  103588.6   31117.4
#> Thomes Creek                  74386.6    42861.0    3265.6  100228.8   25779.4
#> Upper-mid Sacramento River        0.0        0.0       0.0       0.0       0.0
#> Sutter Bypass                     0.0        0.0       0.0       0.0       0.0
#> Bear River                    29798.6    47319.8   10644.6  106414.6  110025.6
#> Feather River               2861544.8 13644336.2 2020621.4  408294.2 9058617.4
#> Yuba River                 10819780.6  7136466.4 2850523.4 9021094.4 7391560.0
#> Lower-mid Sacramento River        0.0        0.0       0.0       0.0       0.0
#> Yolo Bypass                       0.0        0.0       0.0       0.0       0.0
#> American River              7092004.0  4773647.8 2199099.0 5485297.4 5011220.2
#> Lower Sacramento River            0.0        0.0       0.0       0.0       0.0
#> Calaveras River               35387.8    31933.8   37868.4  238357.4    9608.4
#> Cosumnes River                44745.0     7096.4   21603.2  177190.2   60382.2
#> Mokelumne River                 847.8     1350.2   38904.6  143466.6  111061.8
#> Merced River                   8666.4    32844.4  608406.4  407886.0  103714.2
#> Stanislaus River                282.6      251.2  384524.4     910.6 1471561.0
#> Tuolumne River                   94.2      816.4   59785.6     910.6  220647.8
#> San Joaquin River                 0.0        0.0       0.0       0.0       0.0
#> 
#> $proportion_natural
#>                                     1          2           3           4
#> Upper Sacramento River     0.71759765 0.74345801 0.573476523 0.292338292
#> Antelope Creek             0.87431694 0.89021480 0.848684211 0.745856354
#> Battle Creek               0.01139293 0.01126550 0.005264932 0.001717855
#> Bear Creek                 0.87788779 0.88719512 0.850806452 0.758169935
#> Big Chico Creek            0.84918033 0.86144578 0.816733068 0.689189189
#> Butte Creek                0.88054406 0.88925439 0.816863101 0.625231911
#> Clear Creek                0.79823578 0.78917357 0.620393812 0.383085714
#> Cottonwood Creek           0.68360414 0.69582505 0.549242424 0.327261307
#> Cow Creek                  0.73404648 0.77327044 0.632892057 0.342752963
#> Deer Creek                 0.76385542 0.78833693 0.664000000 0.424657534
#> Elder Creek                0.88253968 0.89275362 0.863970588 0.808290155
#> Mill Creek                 0.84304511 0.85750853 0.767409471 0.547425474
#> Paynes Creek               0.88500000 0.89150943 0.845637584 0.751351351
#> Stony Creek                0.87987013 0.88855422 0.854901961 0.779761905
#> Thomes Creek               0.87908497 0.88821752 0.853174603 0.768750000
#> Upper-mid Sacramento River 0.00000000 0.00000000 0.000000000 0.000000000
#> Sutter Bypass              0.00000000 0.00000000 0.000000000 0.000000000
#> Bear River                 0.87788779 0.88922156 0.872852234 0.833333333
#> Feather River              0.29075082 0.34086484 0.223207762 0.073080054
#> Yuba River                 0.32442324 0.37835644 0.349649250 0.316358586
#> Lower-mid Sacramento River 0.00000000 0.00000000 0.000000000 0.000000000
#> Yolo Bypass                0.00000000 0.00000000 0.000000000 0.000000000
#> American River             0.14671840 0.16311717 0.138573451 0.126468625
#> Lower Sacramento River     0.00000000 0.00000000 0.000000000 0.000000000
#> Calaveras River            0.78512397 0.80000000 0.846456693 0.870860927
#> Cosumnes River             0.97461300 0.97731046 0.972721224 0.969719350
#> Mokelumne River            0.04630804 0.05016427 0.040954325 0.032685626
#> Merced River               0.07224917 0.12052275 0.209656372 0.243547044
#> Stanislaus River           0.28816003 0.42073618 0.500758342 0.424868957
#> Tuolumne River             0.18176101 0.25784370 0.305021368 0.256996002
#> San Joaquin River          0.00000000 0.00000000 0.000000000 0.000000000
#>                                      5           6          7           8
#> Upper Sacramento River     0.528336279 0.753468162 0.77784715 0.619575871
#> Antelope Creek             0.850162866 0.931750742 0.94417476 0.894736842
#> Battle Creek               0.003869828 0.009810502 0.01116353 0.005161725
#> Bear Creek                 0.877887789 0.945185185 0.95322377 0.905852417
#> Big Chico Creek            0.830258303 0.921768707 0.93456615 0.876344086
#> Butte Creek                0.783029001 0.886580573 0.88397473 0.786919831
#> Clear Creek                0.554914248 0.725460279 0.71646181 0.517259882
#> Cottonwood Creek           0.626829268 0.822491091 0.83586207 0.695997729
#> Cow Creek                  0.517079705 0.719673406 0.74102011 0.547962382
#> Deer Creek                 0.630188679 0.798767967 0.81190019 0.659722222
#> Elder Creek                0.922431866 0.966455122 0.97049442 0.943251534
#> Mill Creek                 0.646934461 0.788874842 0.80936073 0.667992048
#> Paynes Creek               0.870056497 0.940491591 0.94819820 0.897321429
#> Stony Creek                0.908188586 0.957126304 0.96088795 0.929791271
#> Thomes Creek               0.901595745 0.953634085 0.95823928 0.924949290
#> Upper-mid Sacramento River 0.000000000 0.000000000 0.00000000 0.000000000
#> Sutter Bypass              0.000000000 0.000000000 0.00000000 0.000000000
#> Bear River                 0.903896104 0.954878049 0.96300000 0.935087719
#> Feather River              0.148492570 0.358440509 0.41346768 0.280891214
#> Yuba River                 0.370884304 0.558313927 0.60493780 0.445469664
#> Lower-mid Sacramento River 0.000000000 0.000000000 0.00000000 0.000000000
#> Yolo Bypass                0.000000000 0.000000000 0.00000000 0.000000000
#> American River             0.148328940 0.213375413 0.22758397 0.164932993
#> Lower Sacramento River     0.000000000 0.000000000 0.00000000 0.000000000
#> Calaveras River            0.918918919 0.961557417 0.96517857 0.922233300
#> Cosumnes River             0.980382775 0.990366541 0.99155684 0.985471297
#> Mokelumne River            0.048962551 0.097490695 0.10904350 0.056009042
#> Merced River               0.146347193 0.131037781 0.24590164 0.232523759
#> Stanislaus River           0.364951768 0.394357559 0.48986181 0.510835913
#> Tuolumne River             0.178142767 0.243164631 0.36068796 0.322042731
#> San Joaquin River          0.000000000 0.000000000 0.00000000 0.000000000
#>                                      9          10          11           12
#> Upper Sacramento River     0.420042789 0.474813949 0.366143340 0.2981547692
#> Antelope Creek             0.821011673 0.825095057 0.741573034 0.6320000000
#> Battle Creek               0.001847755 0.001769819 0.001093864 0.0009637675
#> Bear Creek                 0.830275229 0.836283186 0.751677852 0.6300000000
#> Big Chico Creek            0.820312500 0.838596491 0.745856354 0.5892857143
#> Butte Creek                0.774049217 0.819158460 0.699404762 0.3377049180
#> Clear Creek                0.331268583 0.344902913 0.248189415 0.1998221168
#> Cottonwood Creek           0.450628366 0.479970867 0.330206379 0.1901701323
#> Cow Creek                  0.284722222 0.301356589 0.194413408 0.1131611316
#> Deer Creek                 0.493103448 0.496575342 0.352422907 0.2343750000
#> Elder Creek                0.928433269 0.944277108 0.900269542 0.7516778523
#> Mill Creek                 0.510263930 0.492401216 0.374531835 0.2893617021
#> Paynes Creek               0.825757576 0.827715356 0.729411765 0.6101694915
#> Stony Creek                0.893678161 0.884735202 0.817733990 0.7131782946
#> Thomes Creek               0.922105263 0.939739414 0.893371758 0.7909604520
#> Upper-mid Sacramento River 0.000000000 0.000000000 0.000000000 0.0000000000
#> Sutter Bypass              0.000000000 0.000000000 0.000000000 0.0000000000
#> Bear River                 0.875000000 0.864468864 0.784883721 0.5842696629
#> Feather River              0.095771599 0.059864661 0.052734663 0.0601737199
#> Yuba River                 0.240875376 0.231849595 0.164956782 0.0886087110
#> Lower-mid Sacramento River 0.000000000 0.000000000 0.000000000 0.0000000000
#> Yolo Bypass                0.000000000 0.000000000 0.000000000 0.0000000000
#> American River             0.129808811 0.137153407 0.113928765 0.1209525729
#> Lower Sacramento River     0.000000000 0.000000000 0.000000000 0.0000000000
#> Calaveras River            0.913140312 0.944915254 0.903465347 0.7931034483
#> Cosumnes River             0.976979225 0.977422907 0.958543984 0.8681672026
#> Mokelumne River            0.016485673 0.007656766 0.005293106 0.0037110669
#> Merced River               0.117318436 0.043180621 0.025214592 0.0106180234
#> Stanislaus River           0.315068493 0.022035157 0.015698978 0.0117588191
#> Tuolumne River             0.176060798 0.064028777 0.045487894 0.0341499629
#> San Joaquin River          0.000000000 0.000000000 0.000000000 0.0000000000
#>                                     13           14          15          16
#> Upper Sacramento River     0.358524584 0.2391178436 0.077014538 0.453608907
#> Antelope Creek             0.678321678 0.5533980583 0.378378378 0.752688172
#> Battle Creek               0.001509945 0.0009637675 0.000573275 0.004645369
#> Bear Creek                 0.689075630 0.5647058824 0.421875000 0.771604938
#> Big Chico Creek            0.661764706 0.5353535354 0.378378378 0.699346405
#> Butte Creek                0.422857143 0.2887323944 0.140425532 0.520190024
#> Clear Creek                0.248189415 0.1599751012 0.062847222 0.385893060
#> Cottonwood Creek           0.271923861 0.1675087447 0.041610738 0.594778661
#> Cow Creek                  0.148760331 0.0907944515 0.043766578 0.414772727
#> Deer Creek                 0.286407767 0.1855955679 0.095384615 0.453531599
#> Elder Creek                0.812182741 0.6991869919 0.455882353 0.866906475
#> Mill Creek                 0.339920949 0.2374429224 0.143589744 0.409893993
#> Paynes Creek               0.661764706 0.5400000000 0.378378378 0.748633880
#> Stony Creek                0.750000000 0.6224489796 0.455882353 0.801075269
#> Thomes Creek               0.855468750 0.7612903226 0.478873239 0.838427948
#> Upper-mid Sacramento River 0.000000000 0.0000000000 0.000000000 0.000000000
#> Sutter Bypass              0.000000000 0.0000000000 0.000000000 0.000000000
#> Bear River                 0.647619048 0.5256410256 0.430769231 0.719696970
#> Feather River              0.066739209 0.0376343684 0.025880607 0.093107192
#> Yuba River                 0.083192905 0.0525640438 0.030663916 0.141445183
#> Lower-mid Sacramento River 0.000000000 0.0000000000 0.000000000 0.000000000
#> Yolo Bypass                0.000000000 0.0000000000 0.000000000 0.000000000
#> American River             0.139423235 0.1305026745 0.134900172 0.151806838
#> Lower Sacramento River     0.000000000 0.0000000000 0.000000000 0.000000000
#> Calaveras River            0.864818024 0.7952755906 0.458333333 0.546511628
#> Cosumnes River             0.849816850 0.7722222222 0.726666667 0.796019900
#> Mokelumne River            0.005029782 0.0031826018 0.002388852 0.007132479
#> Merced River               0.013304371 0.0122315847 0.013572204 0.021803499
#> Stanislaus River           0.015698978 0.0100250627 0.007038713 0.021065675
#> Tuolumne River             0.046187683 0.0298284862 0.021068473 0.061327561
#> San Joaquin River          0.000000000 0.0000000000 0.000000000 0.000000000
#>                                     17         18         19         20
#> Upper Sacramento River     0.625728687 0.64810121 0.69432064 0.65148361
#> Antelope Creek             0.843537415 0.87150838 0.91136802 0.89616253
#> Battle Creek               0.008607543 0.01202957 0.01638756 0.01021935
#> Bear Creek                 0.853174603 0.88141026 0.92043011 0.90864198
#> Big Chico Creek            0.796460177 0.87862797 0.92508143 0.90274841
#> Butte Creek                0.662771285 0.71060172 0.78232759 0.74168798
#> Clear Creek                0.548133266 0.54308448 0.57294304 0.51578759
#> Cottonwood Creek           0.742671792 0.76404494 0.81318681 0.77125160
#> Cow Creek                  0.577126100 0.61069114 0.66743542 0.58202899
#> Deer Creek                 0.605898123 0.66817156 0.74345550 0.68250540
#> Elder Creek                0.922431866 0.94882434 0.96861747 0.96105263
#> Mill Creek                 0.544959128 0.61958998 0.71979866 0.68549906
#> Paynes Creek               0.839721254 0.87078652 0.91187739 0.89686099
#> Stony Creek                0.877887789 0.90339426 0.93924466 0.93439716
#> Thomes Creek               0.901333333 0.91474654 0.94342508 0.93864013
#> Upper-mid Sacramento River 0.000000000 0.00000000 0.00000000 0.00000000
#> Sutter Bypass              0.000000000 0.00000000 0.00000000 0.00000000
#> Bear River                 0.808290155 0.83982684 0.90997567 0.92060086
#> Feather River              0.132946029 0.09322864 0.16296349 0.23753023
#> Yuba River                 0.220143342 0.33788112 0.49122677 0.47762589
#> Lower-mid Sacramento River 0.000000000 0.00000000 0.00000000 0.00000000
#> Yolo Bypass                0.000000000 0.00000000 0.00000000 0.00000000
#> American River             0.154710765 0.16549217 0.20571186 0.19255214
#> Lower Sacramento River     0.000000000 0.00000000 0.00000000 0.00000000
#> Calaveras River            0.640552995 0.73825503 0.85255198 0.87296417
#> Cosumnes River             0.841698842 0.87687688 0.91581109 0.90949227
#> Mokelumne River            0.011051177 0.01144135 0.01815569 0.02389300
#> Merced River               0.028601978 0.02989856 0.04869110 0.11603016
#> Stanislaus River           0.033047736 0.03446590 0.05343877 0.16118072
#> Tuolumne River             0.094011142 0.09778086 0.14632546 0.19492574
#> San Joaquin River          0.000000000 0.00000000 0.00000000 0.00000000
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    0    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    0    1
#>  [8,]    0    0    0    1
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    1    0    0    0
#>