Fall Run Chinook Model

Fall Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

fall_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = fallRunDSM::params,
  stochastic = FALSE
)

Arguments

scenario
Model inputs, can be modified to test management actions

mode
The mode to run model in. Can be "seed", "simulate", "calibrate"

seeds
The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params
Parameters for model and submodels. Defaults to fallRunDSM::params.

stochastic
TRUE FALSE value indicating if model should be run stochastically. Defaults to FALSE.

Source

IP-117068

Examples

fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = fall_run_seeds)

#> $spawners
#>                                 1      2      3      4      5      6      7
#> Upper Sacramento River      26342  25752  21872  19878  22492  27855  27616
#> Antelope Creek                205    217    171    108    157    267    298
#> Battle Creek                37681  37663  37627  37603  37646  37709  37692
#> Bear Creek                    193    206    161     99    150    264    291
#> Big Chico Creek               201    213    166    103    155    263    295
#> Butte Creek                   906    890    770    512    569    874    826
#> Clear Creek                  5508   4821   3800   3374   3714   4461   4171
#> Cottonwood Creek             2771   2740   2648   2447   2729   3246   3112
#> Cow Creek                    1010   1008    891    793    853    992   1016
#> Deer Creek                    502    501    430    360    417    538    559
#> Elder Creek                   197    208    163    105    170    292    314
#> Mill Creek                    585    634    464    295    326    449    473
#> Paynes Creek                  210    222    171    107    165    280    306
#> Stony Creek                   196    209    163    102    165    283    310
#> Thomes Creek                  196    208    161    100    161    279    307
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    215    232    195    151    264    490    552
#> Feather River              166422 172182 152378 134551 141753 169798 177017
#> Yuba River                  31450  32916  30705  29081  33596  46572  48574
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              30929  32240  29424  27922  29301  32590  33789
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River               378    408    925   1477   3587   9976  14135
#> Cosumnes River               1823   2034   1788   1776   2878   6091   7102
#> Mokelumne River              7958   8012   7908   7858   8113   8906   9065
#> Merced River                 3782   4030   4583   4798   4177   4023   4718
#> Stanislaus River             5577   6650   7774   6945   6292   6439   7495
#> Tuolumne River               1469   1591   1700   1620   1482   1541   1739
#> San Joaquin River               0      0      0      0      0      0      0
#>                                 8      9     10     11     12     13     14
#> Upper Sacramento River      23811  22009  21070  20223  21127  22398  20004
#> Antelope Creek                185    135    122    101     96    108     85
#> Battle Creek                37640  37615  37603  37594  37607  37622  37603
#> Bear Creek                    176    125    111     91     86    100     77
#> Big Chico Creek               183    133    122     99     94    108     85
#> Butte Creek                   641    512    379    306    289    315    268
#> Clear Creek                  3543   3297   3101   2997   3125   3283   3025
#> Cottonwood Creek             2771   2500   2282   2223   2264   2338   2233
#> Cow Creek                     871    806    787    765    764    779    756
#> Deer Creek                    435    383    368    347    344    357    332
#> Elder Creek                   194    137    118     96     86    100     77
#> Mill Creek                    342    282    260    234    233    246    215
#> Paynes Creek                  185    134    120    101     94    108     85
#> Stony Creek                   196    138    115     93     88    100     77
#> Thomes Creek                  195    140    120     97     93    108     84
#> Upper-mid Sacramento River      0      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0      0
#> Bear River                    375    298    306    212    170    208    146
#> Feather River              155581 137384 134205 132478 133154 134495 130900
#> Yuba River                  35979  29190  28841  26563  24854  24794  23126
#> Lower-mid Sacramento River      0      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0      0
#> American River              30487  29264  30306  29578  29745  30188  29674
#> Lower Sacramento River          0      0      0      0      0      0      0
#> Calaveras River             12365  17044  21367  15464  12059   8948  10930
#> Cosumnes River               4304   3050   3284   1886    801    615    381
#> Mokelumne River              8221   7611   7479   7442   7416   7421   7412
#> Merced River                 4829   4165   3755   3674   3605   3612   3619
#> Stanislaus River             7877   5734   4074   4000   3938   3938   3923
#> Tuolumne River               1714   1502   1366   1338   1321   1336   1315
#> San Joaquin River               0      0      0      0      0      0      0
#>                                15     16     17     18     19     20
#> Upper Sacramento River      17650  24355  32184  31390  31583  29062
#> Antelope Creek                 80    125    162    155    200    206
#> Battle Creek                37589  37621  37652  37648  37659  37643
#> Bear Creek                     72    114    148    144    191    197
#> Big Chico Creek                80    117    149    152    205    207
#> Butte Creek                   244    388    529    487    531    512
#> Clear Creek                  2791   3546   4348   4001   3797   3618
#> Cottonwood Creek             2154   2783   3406   3086   3006   2946
#> Cow Creek                     742    875   1001    944    947    927
#> Deer Creek                    324    390    448    436    480    474
#> Elder Creek                    72    128    176    167    212    216
#> Mill Creek                    202    266    326    333    398    378
#> Paynes Creek                   80    124    160    153    200    206
#> Stony Creek                    72    119    158    149    197    206
#> Thomes Creek                   72    120    158    148    194    205
#> Upper-mid Sacramento River      0      0      0      0      0      0
#> Sutter Bypass                   0      0      0      0      0      0
#> Bear River                     99    158    205    229    355    376
#> Feather River              129576 137708 143121 137478 143728 152793
#> Yuba River                  22012  25874  28950  31942  40238  39875
#> Lower-mid Sacramento River      0      0      0      0      0      0
#> Yolo Bypass                     0      0      0      0      0      0
#> American River              29872  30135  29497  29952  31719  30937
#> Lower Sacramento River          0      0      0      0      0      0
#> Calaveras River             16223  11710   9863  14999  16626  16082
#> Cosumnes River                359    337    351    465    628    527
#> Mokelumne River              7418   7437   7452   7457   7494   7580
#> Merced River                 3639   3647   3647   3642   3699   4037
#> Stanislaus River             3924   3959   3982   3978   4032   4529
#> Tuolumne River               1312   1347   1372   1366   1416   1492
#> San Joaquin River               0      0      0      0      0      0
#> 
#> $juvenile_biomass
#>                                    1         2          3          4          5
#> Upper Sacramento River      318678.6  533925.6  2271256.2  3470861.8  4036689.8
#> Antelope Creek                1381.6    1915.4     4270.4     5809.0     6688.2
#> Battle Creek                  6499.8   12748.4    87888.6    96774.8   122303.0
#> Bear Creek                    1099.0    1444.4     5086.8     6342.8     7127.8
#> Big Chico Creek                659.4    1004.8     3045.8     4019.2     4835.6
#> Butte Creek                  15511.6   23550.0   161490.2   175745.8   257040.4
#> Clear Creek                  64652.6  105001.6   429803.2   526578.0   624765.8
#> Cottonwood Creek             20472.8   36643.8   247965.8   278486.6   391432.4
#> Cow Creek                     4898.4    5934.6    20661.2    28636.8    31462.8
#> Deer Creek                    5024.0    6311.4    25905.0    31337.2    34320.2
#> Elder Creek                   1978.2    3799.4    13847.4    15323.2    20190.2
#> Mill Creek                   10173.6   14224.2    31117.4    41228.2    46252.2
#> Paynes Creek                  1318.8    1821.2     7410.4     8949.0     9891.0
#> Stony Creek                   1475.8    2449.2    10581.8    12089.0    17521.2
#> Thomes Creek                  1287.4    2198.0     9608.4    10927.2    16076.8
#> Upper-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Sutter Bypass                    0.0       0.0        0.0        0.0        0.0
#> Bear River                   20190.2   26250.4   108549.8   173516.4   225734.6
#> Feather River               833073.4 1500323.4 18138932.2 37788675.4 45474139.4
#> Yuba River                 1817180.8 2276971.0  7538826.0 12468343.4 14328573.6
#> Lower-mid Sacramento River       0.0       0.0        0.0        0.0        0.0
#> Yolo Bypass                      0.0       0.0        0.0        0.0        0.0
#> American River             2711798.2 4632913.0 11397069.6 22478223.8 24721722.4
#> Lower Sacramento River           0.0       0.0        0.0        0.0        0.0
#> Calaveras River             319400.8  342479.8  1659647.0  3758988.2  4985880.4
#> Cosumnes River              359153.2  395357.4  1311923.4  2300897.8  2741125.8
#> Mokelumne River             217350.8  220585.0   713313.8  1188615.6  1348535.8
#> Merced River                381698.4  385968.8   407666.2   587619.6   952424.8
#> Stanislaus River            457906.2 1066155.6  1628498.2  2120913.0  3680927.8
#> Tuolumne River              114672.8  115834.6   121863.4   167676.0   329166.2
#> San Joaquin River                0.0       0.0        0.0        0.0        0.0
#>                                     6          7          8          9
#> Upper Sacramento River      4385543.8  4763945.2  4980354.0  5129535.4
#> Antelope Creek                 7536.0     8164.0     8478.0     8697.8
#> Battle Creek                 128237.6   136464.4   140169.6   144628.4
#> Bear Creek                     7755.8     8289.6     8415.2     8446.6
#> Big Chico Creek                5557.8     6499.8     6625.4     6656.8
#> Butte Creek                  271421.6   295285.6   299147.8   302413.4
#> Clear Creek                  671803.0   709483.0   735953.2   756363.2
#> Cottonwood Creek             402516.6   422016.0   426914.4   433477.0
#> Cow Creek                     33001.4    34445.8    35042.4    35419.2
#> Deer Creek                    36738.0    39312.8    40129.2    40725.8
#> Elder Creek                   20943.8    23487.2    23738.4    23989.6
#> Mill Creek                    51558.8    54510.4    56708.4    57807.4
#> Paynes Creek                  10738.8    11366.8    11618.0    11712.2
#> Stony Creek                   18683.0    19687.8    20033.2    20410.0
#> Thomes Creek                  17458.4    19373.8    20127.4    20410.0
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                   239927.4   320091.6   326434.4   331678.2
#> Feather River              46149082.4 49940224.2 50434711.4 52096619.2
#> Yuba River                 15088767.6 16786188.8 17353367.0 17902019.2
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             26794436.4 32802606.6 34530642.8 37280121.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River             6289294.4 10511966.4 10908831.0 17672485.2
#> Cosumnes River              2892191.2  3567354.0  3597686.4  4005603.8
#> Mokelumne River             1352869.0  1383389.8  1384614.4  1402324.0
#> Merced River                 956475.4  1004988.4  1007469.0  1007845.8
#> Stanislaus River            3711825.4  3747558.6  3763164.4  3763666.8
#> Tuolumne River               330139.6   343453.2   344018.4   344081.2
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    10         11         12         13
#> Upper Sacramento River      5886432.4  5984369.0  6055301.6  6121210.2
#> Antelope Creek                 9639.8     9671.2     9702.6     9734.0
#> Battle Creek                 176530.8   178980.0   184977.4   187018.4
#> Bear Creek                     9106.0     9106.0     9137.4     9168.8
#> Big Chico Creek                7347.6     7347.6     7379.0     7410.4
#> Butte Creek                  317171.4   317893.6   319620.6   321096.4
#> Clear Creek                  865227.0   877190.4   892105.4   900991.6
#> Cottonwood Creek             473041.0   475710.0   477468.4   480671.2
#> Cow Creek                     39030.2    39250.0    39564.0    39752.4
#> Deer Creek                    44996.2    45184.6    45467.2    45655.6
#> Elder Creek                   24931.6    24931.6    24963.0    25088.6
#> Mill Creek                    63993.2    64401.4    64872.4    65186.4
#> Paynes Creek                  12465.8    12497.2    12528.6    12560.0
#> Stony Creek                   21320.6    21383.4    21414.8    21446.2
#> Thomes Creek                  23550.0    23675.6    23832.6    24146.6
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                   372969.2   375387.0   377930.4   383990.6
#> Feather River              54983692.2 55587545.6 55860034.8 57781777.6
#> Yuba River                 18674993.0 19002243.8 19061558.4 19544270.6
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             41669307.2 43445008.6 46904001.2 49837609.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River            18329969.8 18890679.6 22256696.8 22641535.2
#> Cosumnes River              4054242.4  4072705.6  4130356.0  4134563.6
#> Mokelumne River             1404741.8  1406249.0  1418024.0  1419280.0
#> Merced River                1009510.0  1009981.0  1027502.2  1028695.4
#> Stanislaus River            3765488.0  3765864.8  3772929.8  3774342.8
#> Tuolumne River               344489.4   344583.6   346059.4   346342.0
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    14         15         16         17
#> Upper Sacramento River      9180229.6  9331012.4 11708463.4 13144856.4
#> Antelope Creek                15009.2    15072.0    19719.2    21289.2
#> Battle Creek                 239990.2   241089.2   288440.4   296196.2
#> Bear Creek                    13345.0    13376.4    17301.4    18588.8
#> Big Chico Creek                9828.2     9859.6    15762.8    16516.4
#> Butte Creek                  393567.6   394698.0   443054.0   470403.4
#> Clear Creek                 1272893.2  1287368.6  1469959.6  1580864.4
#> Cottonwood Creek             779913.2   782645.0   925012.6  1023671.4
#> Cow Creek                     89364.4    89678.4   113008.6   119539.8
#> Deer Creek                    83429.8    83932.2   110904.8   118597.8
#> Elder Creek                   36047.2    36078.6    44336.8    48450.2
#> Mill Creek                    85565.0    86444.2   113008.6   120544.6
#> Paynes Creek                  17301.4    17332.8    21666.0    23141.8
#> Stony Creek                   27789.0    27851.8    32813.0    36047.2
#> Thomes Creek                  30489.4    30677.8    35105.2    38056.8
#> Upper-mid Sacramento River        0.0        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0        0.0
#> Bear River                   415139.4   416489.6   474171.4   533266.2
#> Feather River              65091006.8 65420173.0 68882180.0 82792756.8
#> Yuba River                 21682359.4 21953843.8 26247888.0 29497128.6
#> Lower-mid Sacramento River        0.0        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0        0.0
#> American River             54049322.4 55725831.2 63309433.6 69318326.0
#> Lower Sacramento River            0.0        0.0        0.0        0.0
#> Calaveras River            24026620.6 25502797.4 28288574.0 29174839.0
#> Cosumnes River              4180187.8  4191115.0  4290213.4  4305222.6
#> Mokelumne River             1426596.2  1431337.6  1443615.0  1455515.6
#> Merced River                1032494.8  1032808.8  1037299.0  1054474.8
#> Stanislaus River            3779147.0  3779523.8  3787216.8  3790670.8
#> Tuolumne River               347315.4   347378.2   348477.2   353281.4
#> San Joaquin River                 0.0        0.0        0.0        0.0
#>                                    18         19         20
#> Upper Sacramento River     13358690.4 15099475.0 15794388.4
#> Antelope Creek                21603.2    24492.0    25496.8
#> Battle Creek                 298802.4   317108.6   321441.8
#> Bear Creek                    18840.0    21352.0    22513.8
#> Big Chico Creek               17081.6    19154.0    19844.8
#> Butte Creek                  474391.2   496905.0   526421.0
#> Clear Creek                 1591760.2  1688849.0  1788010.2
#> Cottonwood Creek            1034567.2  1087225.0  1121796.4
#> Cow Creek                    120984.2   132665.0   136621.4
#> Deer Creek                   120293.4   131660.2   137689.0
#> Elder Creek                   49046.8    53725.4    55044.2
#> Mill Creek                   122083.2   137092.4   140797.6
#> Paynes Creek                  23455.8    26313.2    27632.0
#> Stony Creek                   36424.0    39721.0    42421.4
#> Thomes Creek                  38402.2    41165.4    43394.8
#> Upper-mid Sacramento River        0.0        0.0        0.0
#> Sutter Bypass                     0.0        0.0        0.0
#> Bear River                   545418.0   633118.2   734351.8
#> Feather River              84543934.8 86951247.2 95298780.2
#> Yuba River                 30569721.2 33428879.6 37444343.0
#> Lower-mid Sacramento River        0.0        0.0        0.0
#> Yolo Bypass                       0.0        0.0        0.0
#> American River             70517523.4 78773305.6 84787724.4
#> Lower Sacramento River            0.0        0.0        0.0
#> Calaveras River            32358296.6 40569428.0 45255469.8
#> Cosumnes River              4345760.0  4657059.6  4764290.6
#> Mokelumne River             1544785.8  1884533.8  2135388.4
#> Merced River                1295187.2  1446284.0  1500731.6
#> Stanislaus River            4260069.4  4287732.8  6025220.4
#> Tuolumne River               426631.8   431938.4   694505.2
#> San Joaquin River                 0.0        0.0        0.0
#> 
#> $proportion_natural
#>                                      1           2           3            4
#> Upper Sacramento River     0.364968491 0.350419385 0.235186540 0.1584666465
#> Antelope Creek             0.780487805 0.792626728 0.736842105 0.5833333333
#> Battle Creek               0.002919243 0.002442716 0.001488293 0.0008509959
#> Bear Creek                 0.808290155 0.820388350 0.770186335 0.6262626263
#> Big Chico Creek            0.776119403 0.788732394 0.728915663 0.5631067961
#> Butte Creek                0.781456954 0.777528090 0.742857143 0.6132812500
#> Clear Creek                0.520152505 0.451773491 0.304473684 0.2166567872
#> Cottonwood Creek           0.242872609 0.234306569 0.207703927 0.1426236208
#> Cow Creek                  0.300990099 0.299603175 0.207631874 0.1097099622
#> Deer Creek                 0.426294821 0.425149701 0.330232558 0.2000000000
#> Elder Creek                0.812182741 0.822115385 0.773006135 0.6476190476
#> Mill Creek                 0.721367521 0.742902208 0.648706897 0.4474576271
#> Paynes Creek               0.785714286 0.797297297 0.736842105 0.5794392523
#> Stony Creek                0.811224490 0.822966507 0.773006135 0.6372549020
#> Thomes Creek               0.811224490 0.822115385 0.770186335 0.6300000000
#> Upper-mid Sacramento River          NA          NA          NA           NA
#> Sutter Bypass                       NA          NA          NA           NA
#> Bear River                 0.827906977 0.840517241 0.810256410 0.7549668874
#> Feather River              0.242017281 0.267374058 0.172157398 0.0624744521
#> Yuba River                 0.355421304 0.384129299 0.339781794 0.3029125546
#> Lower-mid Sacramento River          NA          NA          NA           NA
#> Yolo Bypass                         NA          NA          NA           NA
#> American River             0.175919040 0.209429280 0.133768352 0.0871714061
#> Lower Sacramento River              NA          NA          NA           NA
#> Calaveras River            0.796296296 0.811274510 0.916756757 0.9478672986
#> Cosumnes River             0.978606692 0.980825959 0.978187919 0.9780405405
#> Mokelumne River            0.073636592 0.079880180 0.067779464 0.0618477984
#> Merced River               0.057641460 0.115632754 0.222343443 0.2571904960
#> Stanislaus River           0.305361305 0.417443609 0.501672241 0.4421886249
#> Tuolumne River             0.131381892 0.197988686 0.249411765 0.2123456790
#> San Joaquin River                   NA          NA          NA           NA
#>                                      5           6          7           8
#> Upper Sacramento River     0.256268896 0.399461497 0.39426419 0.297467557
#> Antelope Creek             0.713375796 0.831460674 0.84899329 0.756756757
#> Battle Creek               0.001992244 0.003659604 0.00321023 0.001833156
#> Bear Creek                 0.753333333 0.859848485 0.87285223 0.789772727
#> Big Chico Creek            0.709677419 0.828897338 0.84745763 0.754098361
#> Butte Creek                0.652021090 0.773455378 0.76029056 0.691107644
#> Clear Creek                0.288368336 0.407531944 0.36633901 0.254022015
#> Cottonwood Creek           0.231220227 0.353666051 0.32583548 0.242872609
#> Cow Creek                  0.172332943 0.288306452 0.30511811 0.189437428
#> Deer Creek                 0.309352518 0.464684015 0.48479428 0.337931034
#> Elder Creek                0.782352941 0.873287671 0.88216561 0.809278351
#> Mill Creek                 0.500000000 0.636971047 0.65539112 0.523391813
#> Paynes Creek               0.727272727 0.839285714 0.85294118 0.756756757
#> Stony Creek                0.775757576 0.869257951 0.88064516 0.811224490
#> Thomes Creek               0.770186335 0.867383513 0.87947883 0.810256410
#> Upper-mid Sacramento River          NA          NA         NA          NA
#> Sutter Bypass                       NA          NA         NA          NA
#> Bear River                 0.859848485 0.924489796 0.93297101 0.901333333
#> Feather River              0.110107017 0.257087834 0.28738483 0.189200481
#> Yuba River                 0.396594833 0.564716997 0.58265739 0.436560216
#> Lower-mid Sacramento River          NA          NA         NA          NA
#> Yolo Bypass                         NA          NA         NA          NA
#> American River             0.130132077 0.217919607 0.24567167 0.163971529
#> Lower Sacramento River              NA          NA         NA          NA
#> Calaveras River            0.978533594 0.992281476 0.99455253 0.993772746
#> Cosumnes River             0.986448923 0.993597110 0.99450859 0.990938662
#> Mokelumne River            0.091334895 0.172243431 0.18676227 0.103272108
#> Merced River               0.146756045 0.114093960 0.24459517 0.261958998
#> Stanislaus River           0.384297521 0.398353782 0.48312208 0.508188397
#> Tuolumne River             0.139001350 0.171966256 0.26624497 0.255542590
#> San Joaquin River                   NA          NA         NA          NA
#>                                      9           10           11           12
#> Upper Sacramento River     0.239947294 0.2060749881 0.1728230233 0.2082169735
#> Antelope Creek             0.666666667 0.6311475410 0.5544554455 0.5312500000
#> Battle Creek               0.001169746 0.0008509959 0.0006117998 0.0009572686
#> Bear Creek                 0.704000000 0.6666666667 0.5934065934 0.5697674419
#> Big Chico Creek            0.661654135 0.6311475410 0.5454545455 0.5212765957
#> Butte Creek                0.613281250 0.4775725594 0.3529411765 0.3148788927
#> Clear Creek                0.198362147 0.1476942922 0.1181181181 0.1542400000
#> Cottonwood Creek           0.160800000 0.0806310254 0.0562303194 0.0733215548
#> Cow Creek                  0.124069479 0.1029224905 0.0771241830 0.0759162304
#> Deer Creek                 0.248041775 0.2173913043 0.1700288184 0.1627906977
#> Elder Creek                0.729927007 0.6864406780 0.6145833333 0.5697674419
#> Mill Creek                 0.421985816 0.3730769231 0.3034188034 0.3004291845
#> Paynes Creek               0.664179104 0.6250000000 0.5544554455 0.5212765957
#> Stony Creek                0.731884058 0.6782608696 0.6021505376 0.5795454545
#> Thomes Creek               0.735714286 0.6916666667 0.6185567010 0.6021505376
#> Upper-mid Sacramento River          NA           NA           NA           NA
#> Sutter Bypass                       NA           NA           NA           NA
#> Bear River                 0.875838926 0.8790849673 0.8254716981 0.7823529412
#> Feather River              0.081807197 0.0600573749 0.0478041637 0.0526382985
#> Yuba River                 0.305515588 0.2971117506 0.2368331890 0.1843566428
#> Lower-mid Sacramento River          NA           NA           NA           NA
#> Yolo Bypass                         NA           NA           NA           NA
#> American River             0.129032258 0.1589784201 0.1382784502 0.1431164902
#> Lower Sacramento River              NA           NA           NA           NA
#> Calaveras River            0.995482281 0.9963963121 0.9950206932 0.9936147276
#> Cosumnes River             0.987213115 0.9881242387 0.9793213150 0.9513108614
#> Mokelumne River            0.031401918 0.0143067255 0.0094060736 0.0059331176
#> Merced River               0.144297719 0.0508655126 0.0299401198 0.0113730929
#> Stanislaus River           0.324380886 0.0490918017 0.0315000000 0.0162519045
#> Tuolumne River             0.150466045 0.0658857980 0.0463378176 0.0340651022
#> San Joaquin River                   NA           NA           NA           NA
#>                                     13           14           15          16
#> Upper Sacramento River     0.253147602 0.1637672466 0.0522379603 0.313159515
#> Antelope Creek             0.583333333 0.4705882353 0.4375000000 0.640000000
#> Battle Creek               0.001355590 0.0008509959 0.0004788635 0.001329045
#> Bear Creek                 0.630000000 0.5194805195 0.4861111111 0.675438596
#> Big Chico Creek            0.583333333 0.4705882353 0.4375000000 0.615384615
#> Butte Creek                0.371428571 0.2611940299 0.1885245902 0.489690722
#> Clear Creek                0.194943649 0.1262809917 0.0530275887 0.254653130
#> Cottonwood Creek           0.102651839 0.0604567846 0.0259981430 0.246137262
#> Cow Creek                  0.093709884 0.0661375661 0.0485175202 0.193142857
#> Deer Creek                 0.193277311 0.1325301205 0.1111111111 0.261538462
#> Elder Creek                0.630000000 0.5194805195 0.4861111111 0.710937500
#> Mill Creek                 0.337398374 0.2418604651 0.1930693069 0.387218045
#> Paynes Creek               0.583333333 0.4705882353 0.4375000000 0.637096774
#> Stony Creek                0.630000000 0.5194805195 0.4861111111 0.689075630
#> Thomes Creek               0.657407407 0.5595238095 0.4861111111 0.691666667
#> Upper-mid Sacramento River          NA           NA           NA          NA
#> Sutter Bypass                       NA           NA           NA          NA
#> Bear River                 0.822115385 0.7465753425 0.6262626263 0.765822785
#> Feather River              0.062084092 0.0363254393 0.0264786689 0.083967525
#> Yuba River                 0.182382835 0.1234108795 0.0790477921 0.216510783
#> Lower-mid Sacramento River          NA           NA           NA          NA
#> Yolo Bypass                         NA           NA           NA          NA
#> American River             0.155691003 0.1410662533 0.1467595072 0.154206073
#> Lower Sacramento River              NA           NA           NA          NA
#> Calaveras River            0.991394725 0.9929551693 0.9952536522 0.993424424
#> Cosumnes River             0.936585366 0.8976377953 0.8913649025 0.884272997
#> Mokelumne River            0.006602884 0.0053966541 0.0062011324 0.008740083
#> Merced River               0.013289037 0.0151975684 0.0206100577 0.022758432
#> Stanislaus River           0.016251905 0.0124904410 0.0127420999 0.021470068
#> Tuolumne River             0.044910180 0.0296577947 0.0274390244 0.052709725
#> San Joaquin River                   NA           NA           NA          NA
#>                                    17          18          19          20
#> Upper Sacramento River     0.48023863 0.467091430 0.470347972 0.424403000
#> Antelope Creek             0.72222222 0.709677419 0.775000000 0.781553398
#> Battle Creek               0.00215128 0.002045261 0.002336759 0.001912706
#> Bear Creek                 0.75000000 0.743055556 0.806282723 0.812182741
#> Big Chico Creek            0.69798658 0.703947368 0.780487805 0.782608696
#> Butte Creek                0.62570888 0.593429158 0.627118644 0.613281250
#> Clear Creek                0.39213431 0.339415146 0.303924151 0.269485904
#> Cottonwood Creek           0.38402819 0.320155541 0.302062542 0.287847929
#> Cow Creek                  0.29470529 0.252118644 0.254487856 0.238403452
#> Deer Creek                 0.35714286 0.339449541 0.400000000 0.392405063
#> Elder Creek                0.78977273 0.778443114 0.825471698 0.828703704
#> Mill Creek                 0.50000000 0.510510511 0.590452261 0.568783069
#> Paynes Creek               0.71875000 0.705882353 0.775000000 0.781553398
#> Stony Creek                0.76582278 0.751677852 0.812182741 0.820388350
#> Thomes Creek               0.76582278 0.750000000 0.809278351 0.819512195
#> Upper-mid Sacramento River         NA          NA          NA          NA
#> Sutter Bypass                      NA          NA          NA          NA
#> Bear River                 0.81951220 0.838427948 0.895774648 0.901595745
#> Feather River              0.11861292 0.082435008 0.122335244 0.174405896
#> Yuba River                 0.29975820 0.365349696 0.496197624 0.491611285
#> Lower-mid Sacramento River         NA          NA          NA          NA
#> Yolo Bypass                        NA          NA          NA          NA
#> American River             0.13591213 0.149038462 0.196443772 0.176132140
#> Lower Sacramento River             NA          NA          NA          NA
#> Calaveras River            0.99219304 0.994866324 0.995368700 0.995212038
#> Cosumnes River             0.88888889 0.916129032 0.937898089 0.925996205
#> Mokelumne River            0.01073537 0.011398686 0.016279690 0.027440633
#> Merced River               0.02275843 0.021416804 0.036496350 0.117166213
#> Stanislaus River           0.02712205 0.026143791 0.039186508 0.144623537
#> Tuolumne River             0.06997085 0.065885798 0.098870056 0.144772118
#> San Joaquin River                  NA          NA          NA          NA
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    0    1    0
#>  [3,]    0    0    1    0
#>  [4,]    0    0    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    1    1
#>  [8,]    0    0    1    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    1    1
#> [11,]    0    0    0    1
#> [12,]    0    0    0    1
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    1
#> [16,]    0    0    0    1
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    1
#>