Get Spawning Adults

Calculates the number of initial adults, initial natural adults, and initial adults by month.

get_spawning_adults(
  year,
  adults,
  hatch_adults,
  mode,
  month_return_proportions,
  prop_flow_natal,
  south_delta_routed_watersheds,
  cc_gates_days_closed,
  gates_overtopped,
  tisdale_bypass_watershed,
  yolo_bypass_watershed,
  migratory_temperature_proportion_over_20,
  natural_adult_removal_rate,
  cross_channel_stray_rate,
  stray_rate,
  ..surv_adult_enroute_int,
  .adult_stray_intercept,
  .adult_stray_wild,
  .adult_stray_natal_flow,
  .adult_stray_cross_channel_gates_closed,
  .adult_stray_prop_bay_trans,
  .adult_stray_prop_delta_trans,
  .adult_en_route_migratory_temp,
  .adult_en_route_bypass_overtopped,
  .adult_en_route_adult_harvest_rate,
  stochastic
)

Source

IP-117068

Arguments

year

The year of simulation

adults

Potential spawning adults for each watershed (length = 31) values must be integer

hatch_adults

Total hatchery adults

mode

A value meant to be inherited to determine if model is in "seed", "calibrate", or "simulate" mode

month_return_proportions

The proportion of fish returning for each month

prop_flow_natal

A dataset containing the proportion flows at tributary junction coming from natal watershed using October CALSIM II flows. These proportions are used to estimate straying in the CVPIA SIT Salmon Population Model. More details at proportion_flow_natal

south_delta_routed_watersheds

Indicator of whether fish have access to a region. More details at south_delta_routed_watersheds

cc_gates_days_closed

The number of days and proportion of days the Delta Cross Channel gates are typically closed for each month. More details at delta_cross_channel_closed

gates_overtopped

A monthly TRUE or FALSE value to describe if the gates are overtopped to the bypasses (years 1980-2000) for use with the CVPIA SIT Salmon Population Model to apportion fish onto the bypasses. More details at gates_overtopped

tisdale_bypass_watershed

Indicator of whether fish have access to a region. More details at tisdale_bypass_watershed

yolo_bypass_watershed

Indicator of whether fish have access to a region, More details at yolo_bypass_watershed

migratory_temperature_proportion_over_20

The median proportion of days over 20°C per month. More details at migratory_temperature_proportion_over_20

natural_adult_removal_rate

Spawners removed for hatcheries. More details at natural_adult_removal_rate

cross_channel_stray_rate

Natural straying allocation across channels. More details at cross_channel_stray_rate

stray_rate

Natural straying allocation. More details at stray_rate

..surv_adult_enroute_int

Intercept for surv_adult_enroute

.adult_stray_intercept

Intercept for adult_stray

.adult_stray_wild

Coefficient for adult_stray wild variable

.adult_stray_natal_flow

Coefficient for adult_stray natal_flow variable

.adult_stray_cross_channel_gates_closed

Coefficient for adult_stray cross_channel_gates_closed variable

.adult_stray_prop_bay_trans

Coefficient for adult_stray prop_bay_trans variable

.adult_stray_prop_delta_trans

Coefficient for adult_stray prop_delta_trans variable

.adult_en_route_migratory_temp

Coefficient for surv_adult_enroute migratory_temp variable

.adult_en_route_bypass_overtopped

Coefficient for surv_adult_enroute bypass_overtopped variable

.adult_en_route_adult_harvest_rate

Adult harvest rate for surv_adult_enroute

stochastic

TRUE FALSE value indicating if model is being run stochastically

Details

See params for details on parameter sources